@@ -1,43 +1,95 @@
 [package]
 name = "dioxus-desktop"
-version = "0.1.6"
+version = { workspace = true }
 authors = ["Jonathan Kelley"]
-edition = "2018"
-description = "Dioxus VirtualDOM renderer for a remote webview instance"
-license = "MIT/Apache-2.0"
+edition = "2021"
+description = "WebView renderer for Dioxus"
+license = "MIT OR Apache-2.0"
 repository = "https://github.com/DioxusLabs/dioxus/"
-homepage = "https://dioxuslabs.com"
-documentation = "https://dioxuslabs.com"
-keywords = ["dom", "ui", "gui", "react", "wasm"]
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+homepage = "https://dioxuslabs.com/learn/0.4/getting_started/desktop"
+keywords = ["dom", "ui", "gui", "react"]
 
 [dependencies]
-dioxus-core = { path = "../core", version = "^0.1.9", features = ["serialize"] }
-argh = "0.1.4"
-serde = "1.0.120"
-serde_json = "1.0.61"
-thiserror = "1.0.23"
-log = "0.4.13"
-html-escape = "0.2.9"
-wry = "0.12.2"
-futures-channel = "0.3"
-tokio = { version = "1.12.0", features = [
+dioxus-core = { workspace = true, features = ["serialize"] }
+dioxus-html = { workspace = true, features = [
+    "serialize",
+    "native-bind",
+    "mounted",
+    "eval",
+] }
+dioxus-interpreter-js = { workspace = true, features = ["binary-protocol"] }
+dioxus-hot-reload = { workspace = true, optional = true }
+dioxus-cli-config = { workspace = true }
+generational-box = { workspace = true }
+
+serde = "1.0.136"
+serde_json = "1.0.79"
+thiserror = { workspace = true }
+tracing = { workspace = true }
+wry = { version = "0.35.0", default-features = false, features = [
+    "os-webview",
+    "protocol",
+    "file-drop",
+] }
+futures-channel = { workspace = true }
+tokio = { workspace = true, features = [
     "sync",
     "rt-multi-thread",
     "rt",
     "time",
-], optional = true, default-features = false }
-dioxus-core-macro = { path = "../core-macro", version = "^0.1.7" }
-dioxus-html = { path = "../html", features = ["serialize"], version = "^0.1.6" }
-webbrowser = "0.5.5"
-mime_guess = "2.0.3"
-dioxus-interpreter-js = { path = "../interpreter", version = "^0.0.0" }
+    "macros",
+    "fs",
+], optional = true }
+webbrowser = "0.8.0"
+infer = "0.11.0"
+dunce = "1.0.2"
+slab = { workspace = true }
+rustc-hash = { workspace = true }
+dioxus-hooks = { workspace = true }
 
+futures-util = { workspace = true }
+urlencoding = "2.1.2"
+async-trait = "0.1.68"
+crossbeam-channel = "0.5.8"
+tao = { version = "0.24.0", features = ["rwh_05"] }
+
+[target.'cfg(any(target_os = "windows",target_os = "macos",target_os = "linux",target_os = "dragonfly", target_os = "freebsd", target_os = "netbsd", target_os = "openbsd"))'.dependencies]
+rfd = "0.12"
+global-hotkey = "0.4.1"
+muda = "0.11.3"
+
+[target.'cfg(target_os = "ios")'.dependencies]
+objc = "0.2.7"
+objc_id = "0.1.1"
+
+[target.'cfg(target_os = "macos")'.dependencies]
+core-foundation = "0.9.3"
+objc = "0.2.7"
 
 [features]
-default = ["tokio_runtime"]
+default = ["tokio_runtime", "hot-reload", "wry/objc-exception"]
 tokio_runtime = ["tokio"]
+fullscreen = ["wry/fullscreen"]
+transparent = ["wry/transparent"]
+devtools = ["wry/devtools"]
+hot-reload = ["dioxus-hot-reload"]
+gnu = []
 
+[package.metadata.docs.rs]
+default-features = false
+features = ["tokio_runtime", "hot-reload"]
 
 [dev-dependencies]
-dioxus-hooks = { path = "../hooks" }
+dioxus = { workspace = true, features = ["desktop"] }
+exitcode = "1.1.2"
+
+# These tests need to be run on the main thread, so they cannot use rust's test harness.
+[[test]]
+name = "check_events"
+path = "headless_tests/events.rs"
+harness = false
+
+[[test]]
+name = "check_rendering"
+path = "headless_tests/rendering.rs"
+harness = false
