@@ -1,15 +1,31 @@
 [package]
 name = "dioxus-hooks"
-version = "0.1.7"
+version = { workspace = true }
 authors = ["Jonathan Kelley"]
-edition = "2018"
-description = "Dioxus VirtualDOM renderer for a remote webview instance"
-license = "MIT/Apache-2.0"
+edition = "2021"
+description = "Basic useful hooks for Dioxus."
+license = "MIT OR Apache-2.0"
 repository = "https://github.com/DioxusLabs/dioxus/"
 homepage = "https://dioxuslabs.com"
-documentation = "https://dioxuslabs.com"
-keywords = ["dom", "ui", "gui", "react", "wasm"]
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+keywords = ["dom", "ui", "gui", "react"]
+
+[features]
+default = []
+nightly-features = []
 
 [dependencies]
-dioxus-core = { path = "../../packages/core", version = "^0.1.9" }
+dioxus-core = { workspace = true }
+dioxus-signals = { workspace = true }
+futures-channel = { workspace = true }
+tracing = { workspace = true }
+thiserror = { workspace = true }
+slab = { workspace = true }
+dioxus-debug-cell = "0.1.1"
+futures-util = { workspace = true}
+
+[dev-dependencies]
+futures-util = { workspace = true, default-features = false }
+dioxus-core = { workspace = true }
+dioxus = { workspace = true }
+web-sys = { version = "0.3.64", features = ["Document", "Window", "Element"] }
+tokio = { version = "1.0", features = ["full"] }
