@@ -1,20 +1,79 @@
 [package]
 name = "dioxus-html"
-version = "0.1.6"
+version = { workspace = true }
 authors = ["Jonathan Kelley"]
-edition = "2018"
+edition = "2021"
 description = "HTML Element pack for Dioxus - a concurrent renderer-agnostic Virtual DOM for interactive user experiences"
-license = "MIT/Apache-2.0"
+license = "MIT OR Apache-2.0"
 repository = "https://github.com/DioxusLabs/dioxus/"
 homepage = "https://dioxuslabs.com"
-documentation = "https://docs.rs/dioxus"
-keywords = ["dom", "ui", "gui", "react", "wasm"]
+keywords = ["dom", "ui", "gui", "react"]
 
 [dependencies]
-dioxus-core = { path = "../core", version = "^0.1.9" }
+dioxus-core = { workspace = true }
+dioxus-rsx = { workspace = true, features = ["hot_reload"], optional = true }
+dioxus-html-internal-macro = { workspace = true }
+generational-box = { workspace = true }
 serde = { version = "1", features = ["derive"], optional = true }
 serde_repr = { version = "0.1", optional = true }
+wasm-bindgen = { workspace = true, optional = true }
+euclid = "0.22.7"
+enumset = "1.1.2"
+keyboard-types = "0.7"
+async-trait = "0.1.58"
+serde-value = { version = "0.7.0", optional = true }
+tokio = { workspace = true, features = ["fs", "io-util"], optional = true }
+rfd = { version = "0.12", optional = true }
+futures-channel = { workspace = true }
+serde_json = { version = "1", optional = true }
+
+[dependencies.web-sys]
+optional = true
+version = "0.3.56"
+features = [
+    "Touch",
+    "TouchList",
+    "TouchEvent",
+    "MouseEvent",
+    "DragEvent",
+    "InputEvent",
+    "ClipboardEvent",
+    "KeyboardEvent",
+    "WheelEvent",
+    "AnimationEvent",
+    "TransitionEvent",
+    "PointerEvent",
+    "FocusEvent",
+    "CompositionEvent",
+]
+
+[dev-dependencies]
+serde_json = "1"
 
 [features]
-default = []
-serialize = ["serde", "serde_repr"]
+default = ["serialize", "mounted", "eval"]
+serialize = [
+    "serde",
+    "serde_json",
+    "serde_repr",
+    "euclid/serde",
+    "keyboard-types/serde",
+    "dioxus-core/serialize",
+    "serde-value",
+]
+mounted = [
+    "web-sys?/Element",
+    "web-sys?/DomRect",
+    "web-sys?/ScrollIntoViewOptions",
+    "web-sys?/ScrollLogicalPosition",
+    "web-sys?/ScrollBehavior",
+    "web-sys?/HtmlElement",
+]
+eval = [
+    "serde",
+    "serde_json"
+]
+wasm-bind = ["web-sys", "wasm-bindgen"]
+native-bind = ["tokio"]
+hot-reload-context = ["dioxus-rsx"]
+html-to-rsx = []
