@@ -1,18 +1,16 @@
 [package]
 name = "dioxus-mobile"
-version = "0.0.3"
+version = { workspace = true }
 authors = ["Jonathan Kelley"]
-edition = "2018"
+edition = "2021"
 description = "Mobile-compatible renderer for Dioxus"
 repository = "https://github.com/DioxusLabs/dioxus/"
-homepage = "https://dioxuslabs.com"
-documentation = "https://dioxuslabs.com"
-keywords = ["dom", "ui", "gui", "react", "wasm"]
-license = "MIT/Apache-2.0"
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+homepage = "https://dioxuslabs.com/learn/0.4/getting_started/mobile"
+keywords = ["dom", "ui", "gui", "react"]
+license = "MIT OR Apache-2.0"
 
 [dependencies]
-dioxus-desktop = { path = "../desktop", version = "^0.1.6" }
+dioxus-desktop = { workspace = true, features = ["tokio_runtime"] }
 
 [lib]
 doctest = false
