@@ -1,60 +1,65 @@
 [package]
 name = "dioxus-router"
-version = "0.1.1"
-edition = "2018"
-description = "Dioxus VirtualDOM renderer for the web browser using websys"
-license = "MIT/Apache-2.0"
+authors = ["Jonathan Kelley", "Evan Almloff"]
+version = { workspace = true }
+edition = "2021"
+description = "Cross-platform router for Dioxus apps"
+license = "MIT OR Apache-2.0"
 repository = "https://github.com/DioxusLabs/dioxus/"
 homepage = "https://dioxuslabs.com"
-documentation = "https://dioxuslabs.com"
 keywords = ["dom", "ui", "gui", "react", "wasm"]
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-dioxus-core = { path = "../core", version = "^0.1.9", default-features = false }
-dioxus-html = { path = "../html", version = "^0.1.6", default-features = false }
-dioxus-core-macro = { path = "../core-macro", version = "^0.1.7" }
-
-serde = "1"
-url = "2.2.2"
-serde_urlencoded = "0.7"
-
-# for wasm
-web-sys = { version = "0.3", features = [
-    "Attr",
-    "Document",
-    "History",
-    "HtmlBaseElement",
-    "Event",
-    "NamedNodeMap",
-    "Url",
-    "UrlSearchParams",
-    "Window",
-], optional = true }
-wasm-bindgen = { version = "0.2", optional = true }
-js-sys = { version = "0.3", optional = true }
-gloo = { version = "0.5", optional = true }
-log = "0.4.14"
-
+dioxus-lib = { workspace = true }
+dioxus-router-macro = { workspace = true }
+gloo = { version = "0.8.0", optional = true }
+tracing = { workspace = true }
+urlencoding = "2.1.3"
+serde = { version = "1", features = ["derive"], optional = true }
+serde_json = { version = "1.0.91", optional = true }
+url = "2.3.1"
+wasm-bindgen = { workspace = true, optional = true }
+web-sys = { version = "0.3.60", optional = true, features = [
+    "ScrollRestoration",
+] }
+js-sys = { version = "0.3.63", optional = true }
+gloo-utils = { version = "0.1.6", optional = true }
+dioxus-liveview = { workspace = true, optional = true }
+dioxus-ssr = { workspace = true, optional = true }
+dioxus-fullstack = { workspace = true, optional = true }
+tokio = { workspace = true, features = ["full"], optional = true }
+dioxus-cli-config = { workspace = true }
 
 [features]
-default = ["derive", "web"]
-web = ["web-sys", "gloo", "js-sys", "wasm-bindgen"]
-desktop = []
-mobile = []
-derive = []
+default = []
+ssr = ["dioxus-ssr/incremental", "tokio", "dioxus-fullstack?/server"]
+liveview = ["dioxus-liveview", "tokio", "dep:serde", "serde_json"]
 wasm_test = []
+web = ["gloo", "web-sys", "wasm-bindgen", "gloo-utils", "js-sys"]
+fullstack = ["dioxus-fullstack"]
 
 [dev-dependencies]
-console_error_panic_hook = "0.1.7"
-dioxus-web = { path = "../web" }
-log = "0.4.14"
-wasm-logger = "0.2.0"
-wasm-bindgen-test = "0.3"
-gloo-utils = "0.1.2"
+axum = { workspace = true, features = ["ws"] }
+dioxus = { workspace = true, features = ["router" ]}
+# dioxus-liveview = { workspace = true, features = ["axum"] }
+dioxus-ssr = { path = "../ssr" }
+criterion = { version = "0.5", features = ["async_tokio", "html_reports"] }
 
-[dev-dependencies.web-sys]
-version = "0.3"
-features = [
-    "Document",
-]
+[[example]]
+name = "static_generation"
+required-features = ["ssr"]
+
+[[bench]]
+name = "incremental"
+required-features = ["ssr"]
+harness = false
+
+# [target.'cfg(not(target_family = "wasm"))'.dev-dependencies]
+# dioxus-desktop = { path = "../desktop" }
+
+[target.'cfg(target_family = "wasm")'.dev-dependencies]
+console_error_panic_hook = "0.1.7"
+dioxus-router = { path = ".", features = ["web"] }
+# dioxus-web = { workspace = true }
+gloo = "0.8.0"
+wasm-bindgen-test = "0.3.33"
