@@ -1,30 +1,45 @@
 [package]
 name = "dioxus-ssr"
-version = "0.1.3"
+version = { workspace = true }
 authors = ["Jonathan Kelley"]
-edition = "2018"
+edition = "2021"
 description = "Dioxus render-to-string"
-license = "MIT/Apache-2.0"
+license = "MIT OR Apache-2.0"
 repository = "https://github.com/DioxusLabs/dioxus/"
-homepage = "https://dioxuslabs.com"
-documentation = "https://dioxuslabs.com"
-keywords = ["dom", "ui", "gui", "react", "wasm"]
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+keywords = ["dom", "ui", "gui", "react", "ssr"]
 
 [dependencies]
-dioxus-core = { path = "../core", version = "^0.1.9", features = ["serialize"] }
+dioxus-core = { workspace = true, features = ["serialize"] }
+dioxus-html = { workspace = true, features = ["eval"]}
+generational-box = { workspace = true }
+askama_escape = "0.10.3"
+thiserror = "1.0.23"
+rustc-hash = "1.1.0"
+lru  = { workspace = true }
+tracing = { workspace = true }
+http = { workspace = true }
+async-trait = { workspace = true }
+serde_json = { workspace = true }
+chrono = { verison = "0.4.34", optional = true }
 
+[target.'cfg(target_arch = "wasm32")'.dependencies]
+tokio = { version = "1.28", features = ["io-util"], optional = true }
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+tokio = { version = "1.28", features = ["fs", "io-util"], optional = true }
 
 [dev-dependencies]
-dioxus-hooks = { path = "../hooks" }
-dioxus-html = { path = "../html" }
-dioxus-core-macro = { path = "../core-macro" }
-thiserror = "1.0.23"
-log = "0.4.13"
+dioxus = { workspace = true }
+dioxus-signals = { workspace = true }
+
+tracing = { workspace = true }
 fern = { version = "0.6.0", features = ["colored"] }
 anyhow = "1.0"
 argh = "0.1.4"
 serde = "1.0.120"
 serde_json = "1.0.61"
 fs_extra = "1.2.0"
+
+[features]
+default = []
+incremental = ["dep:tokio", "chrono"]
