@@ -1,82 +1,83 @@
 [package]
 name = "dioxus-web"
-version = "0.0.5"
+version = { workspace = true }
 authors = ["Jonathan Kelley"]
-edition = "2018"
-description = "Dioxus VirtualDOM renderer for the web browser using websys"
-license = "MIT/Apache-2.0"
+edition = "2021"
+description = "Web renderer for Dioxus using websys"
+license = "MIT OR Apache-2.0"
 repository = "https://github.com/DioxusLabs/dioxus/"
-homepage = "https://dioxuslabs.com"
-documentation = "https://dioxuslabs.com"
+homepage = "https://dioxuslabs.com/learn/0.4/getting_started/wasm"
 keywords = ["dom", "ui", "gui", "react", "wasm"]
 
 [dependencies]
-dioxus-core = { path = "../core", version = "^0.1.9" }
-dioxus-html = { path = "../html", version = "^0.1.6" }
+dioxus-core = { workspace = true, features = ["serialize"] }
+dioxus-html = { workspace = true, features = ["wasm-bind"] }
+dioxus-interpreter-js = { workspace = true, features = [
+    "minimal_bindings",
+    "webonly",
+] }
+generational-box = { workspace = true }
+
 js-sys = "0.3.56"
-wasm-bindgen = { version = "0.2.79", features = ["enable-interning"] }
-lazy_static = "1.4.0"
+wasm-bindgen = { workspace = true }
 wasm-bindgen-futures = "0.4.29"
-log = { version = "0.4.14", features = ["release_max_level_off"] }
-fxhash = "0.2.1"
-wasm-logger = "0.2.0"
+tracing = { workspace = true }
+rustc-hash = { workspace = true }
 console_error_panic_hook = { version = "0.1.7", optional = true }
-wasm-bindgen-test = "0.3.29"
-once_cell = "1.9.0"
-async-channel = "1.6.1"
-anyhow = "1.0.53"
-gloo-timers = { version = "0.2.3", features = ["futures"] }
-futures-util = "0.3.19"
-smallstr = "0.2.0"
-dioxus-interpreter-js = { path = "../interpreter", version = "^0.0.0", features = ["web"] }
-serde-wasm-bindgen = "0.4.2"
+futures-util = { workspace = true, features = ["std", "async-await", "async-await-macro"] }
+futures-channel = { workspace = true }
+serde_json = { version = "1.0" }
+serde = { version = "1.0" }
+serde-wasm-bindgen = { version = "0.5.0", optional = true }
+async-trait = { version = "0.1.58", optional = true }
 
 [dependencies.web-sys]
 version = "0.3.56"
 features = [
-    "Comment",
-    "Attr",
     "Document",
-    "Element",
-    "CssStyleDeclaration",
     "HtmlElement",
     "HtmlInputElement",
     "HtmlSelectElement",
     "HtmlTextAreaElement",
     "HtmlFormElement",
-    "EventTarget",
-    "HtmlCollection",
-    "Node",
-    "NodeList",
     "Text",
     "Window",
-    "Event",
-    "MouseEvent",
-    "InputEvent",
-    "ClipboardEvent",
-    "NamedNodeMap",
-    "KeyboardEvent",
-    "TouchEvent",
-    "WheelEvent",
-    "AnimationEvent",
-    "TransitionEvent",
-    "PointerEvent",
-    "FocusEvent",
-    "CompositionEvent",
-    "ClipboardEvent",
-    "DocumentType",
-    "CharacterData",
-    "SvgElement",
-    "SvgAnimatedString",
-    "HtmlOptionElement",
-    "IdleDeadline",
+    "DataTransfer",
+    "console"
 ]
 
 [features]
-default = ["panic_hook"]
+default = ["panic_hook", "mounted", "file_engine", "hot_reload", "eval"]
 panic_hook = ["console_error_panic_hook"]
+hydrate = [
+    "web-sys/Comment",
+]
+mounted = [
+    "web-sys/Element",
+    "dioxus-html/mounted"
+]
+file_engine = [
+    "web-sys/File",
+    "web-sys/FileList",
+    "web-sys/FileReader",
+    "async-trait"
+]
+hot_reload = [
+    "web-sys/MessageEvent",
+    "web-sys/WebSocket",
+    "web-sys/Location",
+]
+eval = [
+    "dioxus-html/eval",
+    "serde-wasm-bindgen",
+    "async-trait"
+]
 
 [dev-dependencies]
-dioxus-core-macro = { path = "../core-macro" }
+dioxus = { workspace = true }
 wasm-bindgen-test = "0.3.29"
-dioxus-ssr = { path = "../ssr" }
+dioxus-ssr = { workspace = true, default-features = false }
+gloo-timers = "0.2.3"
+gloo-dialogs = "0.1.1"
+dioxus-web = { path = ".", features = ["hydrate"] }
+tracing-wasm = "0.2.1"
