@@ -1,23 +1,24 @@
+//! A simple demonstration of how to set attributes on buttons to disable them.
+//!
+//! This example also showcases the shorthand syntax for attributes, and how signals themselves implement IntoAttribute
+
 use dioxus::prelude::*;
 
 fn main() {
-    dioxus::desktop::launch(app);
+    launch(app);
 }
 
-fn app(cx: Scope) -> Element {
-    let (disabled, set_disabled) = use_state(&cx, || false);
+fn app() -> Element {
+    let mut disabled = use_signal(|| false);
 
-    cx.render(rsx! {
-        div {
-            button {
-                onclick: move |_| set_disabled(!disabled),
-                "click to " [if *disabled {"enable"} else {"disable"} ] " the lower button"
+    rsx! {
+        div { style: "text-align: center; margin: 20px; display: flex; flex-direction: column; align-items: center;",
+            button { onclick: move |_| disabled.toggle(),
+                "click to "
+                if disabled() { "enable" } else { "disable" }
+                " the lower button"
             }
-
-            button {
-                disabled: "{disabled}",
-                "lower button"
-            }
+            button { disabled, "lower button" }
         }
-    })
+    }
 }
