@@ -1,64 +1,73 @@
 //! Example: File Explorer
-//! -------------------------
 //!
 //! This is a fun little desktop application that lets you explore the file system.
 //!
 //! This example is interesting because it's mixing filesystem operations and GUI, which is typically hard for UI to do.
-//!
-//! It also uses `use_ref` to maintain a model, rather than `use_state`. That way,
-//! we dont need to clutter our code with `read` commands.
+//! We store the state entirely in a single signal, making the explorer logic fairly easy to reason about.
 
+use dioxus::desktop::{Config, WindowBuilder};
 use dioxus::prelude::*;
 
 fn main() {
-    dioxus::desktop::launch_cfg(app, |c| c.with_window(|w| w.with_resizable(true)));
+    LaunchBuilder::desktop()
+        .with_cfg(Config::new().with_window(WindowBuilder::new().with_resizable(true)))
+        .launch(app)
 }
 
-fn app(cx: Scope) -> Element {
-    let files = use_ref(&cx, Files::new);
+#[cfg(not(feature = "collect-assets"))]
+const _STYLE: &str = include_str!("../examples/assets/fileexplorer.css");
 
-    rsx!(cx, div {
-        link { href:"https://fonts.googleapis.com/icon?family=Material+Icons", rel:"stylesheet", }
-        style { [include_str!("./assets/fileexplorer.css")] }
-        header {
-            i { class: "material-icons icon-menu", "menu" }
-            h1 { "Files: " [files.read().current()] }
-            span { }
-            i { class: "material-icons", onclick: move |_| files.write().go_up(), "logout" }
-        }
-        main {
-            files.read().path_names.iter().enumerate().map(|(dir_id, path)| {
-                let path_end = path.split('/').last().unwrap_or_else(|| path.as_str());
-                let icon_type = if path_end.contains('.') {
-                    "description"
-                } else {
-                    "folder"
-                };
-                rsx! (
-                    div {
-                        class: "folder",
-                        key: "{path}",
-                        i { class: "material-icons",
-                            onclick: move |_| files.write().enter_dir(dir_id),
-                            "{icon_type}"
-                            p { class: "cooltip", "0 folders / 0 files" }
+#[cfg(feature = "collect-assets")]
+const _STYLE: &str = manganis::mg!(file("./examples/assets/fileexplorer.css"));
+
+fn app() -> Element {
+    let mut files = use_signal(Files::new);
+
+    rsx! {
+        div {
+            link { href:"https://fonts.googleapis.com/icon?family=Material+Icons", rel:"stylesheet" }
+            header {
+                i { class: "material-icons icon-menu", "menu" }
+                h1 { "Files: ", {files.read().current()} }
+                span { }
+                i { class: "material-icons", onclick: move |_| files.write().go_up(), "logout" }
+            }
+            style { "{_STYLE}" }
+            main {
+                for (dir_id, path) in files.read().path_names.iter().enumerate() {
+                    {
+                        let path_end = path.split('/').last().unwrap_or(path.as_str());
+                        rsx! {
+                            div { class: "folder", key: "{path}",
+                                i { class: "material-icons",
+                                    onclick: move |_| files.write().enter_dir(dir_id),
+                                    if path_end.contains('.') {
+                                        "description"
+                                    } else {
+                                        "folder"
+                                    }
+                                    p { class: "cooltip", "0 folders / 0 files" }
+                                }
+                                h1 { "{path_end}" }
+                            }
                         }
-                        h1 { "{path_end}" }
                     }
-                )
-            }),
-            files.read().err.as_ref().map(|err| {
-                rsx! (
+                }
+                if let Some(err) = files.read().err.as_ref() {
                     div {
                         code { "{err}" }
                         button { onclick: move |_| files.write().clear_err(), "x" }
                     }
-                )
-            })
+                }
+            }
         }
-    })
+    }
 }
 
+/// A simple little struct to hold the file explorer state
+///
+/// We don't use any fancy signals or memoization here - Dioxus is so fast that even a file explorer can be done with a
+/// single signal.
 struct Files {
     path_stack: Vec<String>,
     path_names: Vec<String>,
@@ -83,7 +92,7 @@ impl Files {
         let paths = match std::fs::read_dir(cur_path) {
             Ok(e) => e,
             Err(err) => {
-                let err = format!("An error occured: {:?}", err);
+                let err = format!("An error occured: {err:?}");
                 self.err = Some(err);
                 self.path_stack.pop();
                 return;
