@@ -9,26 +9,28 @@
 //! In this example, we pre-render the page to HTML and then pass it into the desktop configuration. This serves as a
 //! proof-of-concept for the hydration feature, but you'll probably only want to use hydration for the web.
 
+use dioxus::desktop::Config;
 use dioxus::prelude::*;
-use dioxus::ssr;
 
 fn main() {
-    let vdom = VirtualDom::new(app);
-    let content = ssr::render_vdom_cfg(&vdom, |f| f.pre_render(true));
+    LaunchBuilder::desktop()
+        .with_cfg(Config::new().with_prerendered({
+            // We build the dom a first time, then pre-render it to HTML
+            let pre_rendered_dom = VirtualDom::prebuilt(app);
 
-    dioxus::desktop::launch_cfg(app, |c| c.with_prerendered(content));
+            // We then launch the app with the pre-rendered HTML
+            dioxus_ssr::pre_render(&pre_rendered_dom)
+        }))
+        .launch(app)
 }
 
-fn app(cx: Scope) -> Element {
-    let (val, set_val) = use_state(&cx, || 0);
+fn app() -> Element {
+    let mut val = use_signal(|| 0);
 
-    cx.render(rsx! {
+    rsx! {
         div {
             h1 { "hello world. Count: {val}" }
-            button {
-                onclick: move |_| set_val(val + 1),
-                "click to increment"
-            }
+            button { onclick: move |_| val += 1, "click to increment" }
         }
-    })
+    }
 }
