@@ -1,5 +1,3 @@
-#![allow(non_snake_case)]
-
 //! Suspense in Dioxus
 //!
 //! Currently, `rsx!` does not accept futures as values. To achieve the functionality
@@ -10,49 +8,53 @@
 //! will cause it to fetch a random dog image from the Dog API. Since the data
 //! is not ready immediately, we render some loading text.
 //!
-//! We can achieve the majoirty of suspense functionality by composing "suspenseful"
+//! We can achieve the majority of suspense functionality by composing "suspenseful"
 //! primitives in our own custom components.
 
+use dioxus::desktop::{Config, LogicalSize, WindowBuilder};
 use dioxus::prelude::*;
 
 fn main() {
-    use dioxus::desktop::tao::dpi::LogicalSize;
-    dioxus::desktop::launch_cfg(app, |cfg| {
-        cfg.with_window(|w| {
-            w.with_title("Doggo Fetcher")
-                .with_inner_size(LogicalSize::new(600.0, 800.0))
-        })
-    });
+    LaunchBuilder::desktop()
+        .with_cfg(
+            Config::new().with_window(
+                WindowBuilder::new()
+                    .with_title("Doggo Fetcher")
+                    .with_inner_size(LogicalSize::new(600.0, 800.0)),
+            ),
+        )
+        .launch(app)
 }
 
-#[derive(serde::Deserialize)]
-struct DogApi {
-    message: String,
-}
-
-fn app(cx: Scope) -> Element {
-    cx.render(rsx! {
+fn app() -> Element {
+    rsx! {
         div {
-            h1 {"Dogs are very important"}
+            h1 { "Dogs are very important" }
             p {
-                "The dog or domestic dog (Canis familiaris[4][5] or Canis lupus familiaris[5])" 
-                "is a domesticated descendant of the wolf which is characterized by an upturning tail." 
+                "The dog or domestic dog (Canis familiaris[4][5] or Canis lupus familiaris[5])"
+                "is a domesticated descendant of the wolf which is characterized by an upturning tail."
                 "The dog derived from an ancient, extinct wolf,[6][7] and the modern grey wolf is the"
                 "dog's nearest living relative.[8] The dog was the first species to be domesticated,[9][8]"
                 "by hunter–gatherers over 15,000 years ago,[7] before the development of agriculture.[1]"
             }
 
             h3 { "Illustrious Dog Photo" }
-            Doggo { }
+            Doggo {}
         }
-    })
+    }
 }
 
 /// This component will re-render when the future has finished
 /// Suspense is achieved my moving the future into only the component that
 /// actually renders the data.
-fn Doggo(cx: Scope) -> Element {
-    let fut = use_future(&cx, || async move {
+#[component]
+fn Doggo() -> Element {
+    let mut fut = use_resource(move || async move {
+        #[derive(serde::Deserialize)]
+        struct DogApi {
+            message: String,
+        }
+
         reqwest::get("https://dog.ceo/api/breeds/image/random/")
             .await
             .unwrap()
@@ -60,21 +62,12 @@ fn Doggo(cx: Scope) -> Element {
             .await
     });
 
-    cx.render(match fut.value() {
+    match fut.read().as_ref() {
         Some(Ok(resp)) => rsx! {
-            button {
-                onclick: move |_| fut.restart(),
-                "Click to fetch another doggo"
-            }
-            div {
-                img {
-                    max_width: "500px",
-                    max_height: "500px",
-                    src: "{resp.message}",
-                }
-            }
+            button { onclick: move |_| fut.restart(), "Click to fetch another doggo" }
+            div { img { max_width: "500px", max_height: "500px", src: "{resp.message}" } }
         },
         Some(Err(_)) => rsx! { div { "loading dogs failed" } },
         None => rsx! { div { "loading dogs..." } },
-    })
+    }
 }
