@@ -1,101 +1,56 @@
-// Thanks to @japsu and their project https://github.com/japsu/jatsi for the example!
+//! Thanks to @japsu and their project https://github.com/japsu/jatsi for the example!
+//!
+//! This example shows how to create a simple dice rolling app using SVG and Dioxus.
+//! The `svg` element and its children have a custom namespace, and are attached using different methods than regular
+//! HTML elements. Any element can specify a custom namespace by using the `namespace` meta attribute.
+//!
+//! If you `go-to-definition` on the `svg` element, you'll see its custom namespace.
 
-use dioxus::{events::MouseEvent, prelude::*};
+use dioxus::prelude::*;
+use rand::{thread_rng, Rng};
 
 fn main() {
-    dioxus::desktop::launch(app);
+    launch(|| {
+        rsx! {
+            div { user_select: "none", webkit_user_select: "none", margin_left: "10%", margin_right: "10%",
+                h1 { "Click die to generate a new value" }
+                div { cursor: "pointer", height: "100%", width: "100%", Dice {} }
+            }
+        }
+    });
 }
 
-fn app(cx: Scope) -> Element {
-    let (val, set_val) = use_state(&cx, || 5);
+#[component]
+fn Dice() -> Element {
+    const Y: bool = true;
+    const N: bool = false;
+    const DOTS: [(i64, i64); 7] = [(-1, -1), (-1, -0), (-1, 1), (1, -1), (1, 0), (1, 1), (0, 0)];
+    const DOTS_FOR_VALUE: [[bool; 7]; 6] = [
+        [N, N, N, N, N, N, Y],
+        [N, N, Y, Y, N, N, N],
+        [N, N, Y, Y, N, N, Y],
+        [Y, N, Y, Y, N, Y, N],
+        [Y, N, Y, Y, N, Y, Y],
+        [Y, Y, Y, Y, Y, Y, N],
+    ];
 
-    cx.render(rsx! {
-        div {
-            user_select: "none",
-            webkit_user_select: "none",
-            margin_left: "10%",
-            margin_right: "10%",
-            h1 { "Click die to generate a new value" }
-            div {
-                cursor: "pointer",
-                height: "80%",
-                width: "80%",
-                Die {
-                    value: *val,
-                    keep: true,
-                    onclick: move |_| {
-                        use rand::Rng;
-                        let mut rng = rand::thread_rng();
-                        set_val(rng.gen_range(1..6));
-                    }
+    let mut value = use_signal(|| 5);
+    let active_dots = use_memo(move || &DOTS_FOR_VALUE[(value() - 1) as usize]);
+
+    rsx! {
+        svg {
+            view_box: "-1000 -1000 2000 2000",
+            prevent_default: "onclick",
+            onclick: move |_| value.set(thread_rng().gen_range(1..=6)),
+            rect { x: -1000, y: -1000, width: 2000, height: 2000, rx: 200, fill: "#aaa" }
+            for ((x, y), _) in DOTS.iter().zip(active_dots.read().iter()).filter(|(_, &active)| active) {
+                circle {
+                    cx: *x * 600,
+                    cy: *y * 600,
+                    r: 200,
+                    fill: "#333"
                 }
             }
         }
-    })
-}
-
-#[derive(Props)]
-pub struct DieProps<'a> {
-    pub value: u64,
-    pub keep: bool,
-    pub onclick: EventHandler<'a, MouseEvent>,
-}
-
-const DOTS: [(i64, i64); 7] = [(-1, -1), (-1, -0), (-1, 1), (1, -1), (1, 0), (1, 1), (0, 0)];
-const DOTS_FOR_VALUE: [[bool; 7]; 6] = [
-    [false, false, false, false, false, false, true],
-    [false, false, true, true, false, false, false],
-    [false, false, true, true, false, false, true],
-    [true, false, true, true, false, true, false],
-    [true, false, true, true, false, true, true],
-    [true, true, true, true, true, true, false],
-];
-
-const OFFSET: i64 = 600;
-const DOT_RADIUS: &str = "200";
-const HELD_COLOR: &str = "#aaa";
-const UNHELD_COLOR: &str = "#ddd";
-
-// A six-sided die (D6) with dots.
-#[allow(non_snake_case)]
-pub fn Die<'a>(cx: Scope<'a, DieProps<'a>>) -> Element {
-    let &DieProps { value, keep, .. } = cx.props;
-
-    let active_dots = &DOTS_FOR_VALUE[(value - 1) as usize];
-    let fill = if keep { HELD_COLOR } else { UNHELD_COLOR };
-    let dots = DOTS
-        .iter()
-        .zip(active_dots.iter())
-        .filter(|(_, &active)| active)
-        .map(|((x, y), _)| {
-            let dcx = x * OFFSET;
-            let dcy = y * OFFSET;
-            rsx!(circle {
-                cx: "{dcx}",
-                cy: "{dcy}",
-                r: "{DOT_RADIUS}",
-                fill: "#333"
-            })
-        });
-
-    rsx!(cx,
-      svg {
-        onclick: move |e| cx.props.onclick.call(e),
-        prevent_default: "onclick",
-        "dioxus-prevent-default": "onclick",
-        class: "die",
-        view_box: "-1000 -1000 2000 2000",
-
-        rect {
-          x: "-1000",
-          y: "-1000",
-          width: "2000",
-          height: "2000",
-          rx: "{DOT_RADIUS}",
-          fill: "{fill}",
-        }
-
-        dots
-      }
-    )
+    }
 }
