@@ -5,23 +5,21 @@
 use dioxus::prelude::*;
 
 fn main() {
-    dioxus::desktop::launch(app);
+    launch_desktop(app);
 }
 
-fn app(cx: Scope) -> Element {
-    let (contents, set_contents) = use_state(&cx, || {
-        String::from("<script>alert(\"hello world\")</script>")
-    });
+fn app() -> Element {
+    let mut contents = use_signal(|| String::from("<script>alert(\"hello world\")</script>"));
 
-    cx.render(rsx! {
+    rsx! {
         div {
             h1 {"Dioxus is XSS-Safe"}
             h3 { "{contents}" }
             input {
                 value: "{contents}",
                 r#type: "text",
-                oninput: move |e| set_contents(e.value.clone()),
+                oninput: move |e| contents.set(e.value()),
             }
         }
-    })
+    }
 }
