@@ -1,24 +1,35 @@
+#![doc = include_str!("../README.md")]
+#![doc(html_logo_url = "https://avatars.githubusercontent.com/u/79236386")]
+#![doc(html_favicon_url = "https://avatars.githubusercontent.com/u/79236386")]
+
 use proc_macro::TokenStream;
 use quote::ToTokens;
-use syn::parse_macro_input;
+use rsx::RenderCallBody;
+use syn::parse::Parser;
+use syn::punctuated::Punctuated;
+use syn::{parse_macro_input, Path, Token};
 
-pub(crate) mod ifmt;
-pub(crate) mod inlineprops;
-pub(crate) mod props;
-pub(crate) mod router;
-pub(crate) mod rsx;
+mod component_body;
+mod component_body_deserializers;
+mod props;
+mod utils;
+
+// mod rsx;
+use crate::component_body::ComponentBody;
+use crate::component_body_deserializers::component::ComponentDeserializerArgs;
+use crate::component_body_deserializers::inline_props::InlinePropsDeserializerArgs;
+use dioxus_rsx as rsx;
 
 #[proc_macro]
 pub fn format_args_f(input: TokenStream) -> TokenStream {
-    use ifmt::*;
-    let item = parse_macro_input!(input as IfmtInput);
-    format_args_f_impl(item)
+    use rsx::*;
+    format_args_f_impl(parse_macro_input!(input as IfmtInput))
         .unwrap_or_else(|err| err.to_compile_error())
         .into()
 }
 
 #[proc_macro_derive(Props, attributes(props))]
-pub fn derive_typed_builder(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
+pub fn derive_typed_builder(input: TokenStream) -> TokenStream {
     let input = parse_macro_input!(input as syn::DeriveInput);
     match props::impl_my_derive(&input) {
         Ok(output) => output.into(),
@@ -27,193 +38,21 @@ pub fn derive_typed_builder(input: proc_macro::TokenStream) -> proc_macro::Token
 }
 
 /// The rsx! macro makes it easy for developers to write jsx-style markup in their components.
-///
-/// ## Complete Reference Guide:
-/// ```
-/// const Example: Component = |cx| {
-///     let formatting = "formatting!";
-///     let formatting_tuple = ("a", "b");
-///     let lazy_fmt = format_args!("lazily formatted text");
-///     cx.render(rsx! {
-///         div {
-///             // Elements
-///             div {}
-///             h1 {"Some text"}
-///             h1 {"Some text with {formatting}"}
-///             h1 {"Formatting basic expressions {formatting_tuple.0} and {formatting_tuple.1}"}
-///             h2 {
-///                 "Multiple"
-///                 "Text"
-///                 "Blocks"
-///                 "Use comments as separators in html"
-///             }
-///             div {
-///                 h1 {"multiple"}
-///                 h2 {"nested"}
-///                 h3 {"elements"}
-///             }
-///             div {
-///                 class: "my special div"
-///                 h1 {"Headers and attributes!"}
-///             }
-///             div {
-///                 // pass simple rust expressions in
-///                 class: lazy_fmt,
-///                 id: format_args!("attributes can be passed lazily with std::fmt::Arguments"),
-///                 div {
-///                     class: {
-///                         const WORD: &str = "expressions";
-///                         format_args!("Arguments can be passed in through curly braces for complex {}", WORD)
-///                     }
-///                 }
-///             }
-///
-///             // Expressions can be used in element position too:
-///             {rsx!(p { "More templating!" })}
-///             {html!(<p>"Even HTML templating!!"</p>)}
-///
-///             // Iterators
-///             {(0..10).map(|i| rsx!(li { "{i}" }))}
-///             {{
-///                 let data = std::collections::HashMap::<&'static str, &'static str>::new();
-///                 // Iterators *should* have keys when you can provide them.
-///                 // Keys make your app run faster. Make sure your keys are stable, unique, and predictable.
-///                 // Using an "ID" associated with your data is a good idea.
-///                 data.into_iter().map(|(k, v)| rsx!(li { key: "{k}" "{v}" }))
-///             }}
-///
-///             // Matching
-///             {match true {
-///                 true => rsx!(h1 {"Top text"}),
-///                 false => rsx!(h1 {"Bottom text"})
-///             }}
-///
-///             // Conditional rendering
-///             // Dioxus conditional rendering is based around None/Some. We have no special syntax for conditionals.
-///             // You can convert a bool condition to rsx! with .then and .or
-///             {true.then(|| rsx!(div {}))}
-///
-///             // True conditions
-///             {if true {
-///                 rsx!(h1 {"Top text"})
-///             } else {
-///                 rsx!(h1 {"Bottom text"})
-///             }}
-///
-///             // returning "None" is a bit noisy... but rare in practice
-///             {None as Option<()>}
-///
-///             // Use the Dioxus type-alias for less noise
-///             {NONE_ELEMENT}
-///
-///             // can also just use empty fragments
-///             Fragment {}
-///
-///             // Fragments let you insert groups of nodes without a parent.
-///             // This lets you make components that insert elements as siblings without a container.
-///             div {"A"}
-///             Fragment {
-///                 div {"B"}
-///                 div {"C"}
-///                 Fragment {
-///                     "D"
-///                     Fragment {
-///                         "heavily nested fragments is an antipattern"
-///                         "they cause Dioxus to do unnecessary work"
-///                         "don't use them carelessly if you can help it"
-///                     }
-///                 }
-///             }
-///
-///             // Components
-///             // Can accept any paths
-///             // Notice how you still get syntax highlighting and IDE support :)
-///             Baller {}
-///             baller::Baller { }
-///             crate::baller::Baller {}
-///
-///             // Can take properties
-///             Taller { a: "asd" }
-///
-///             // Can take optional properties
-///             Taller { a: "asd" }
-///
-///             // Can pass in props directly as an expression
-///             {{
-///                 let props = TallerProps {a: "hello"};
-///                 rsx!(Taller { ..props })
-///             }}
-///
-///             // Spreading can also be overridden manually
-///             Taller {
-///                 ..TallerProps { a: "ballin!" }
-///                 a: "not ballin!"
-///             }
-///
-///             // Can take children too!
-///             Taller { a: "asd", div {"hello world!"} }
-///         }
-///     })
-/// };
-///
-/// mod baller {
-///     use super::*;
-///     pub struct BallerProps {}
-///
-///     /// This component totally balls
-///     pub fn Baller(cx: Scope) -> DomTree {
-///         todo!()
-///     }
-/// }
-///
-/// #[derive(Debug, PartialEq, Props)]
-/// pub struct TallerProps {
-///     a: &'static str,
-/// }
-///
-/// /// This component is taller than most :)
-/// pub fn Taller(cx: Scope<TallerProps>) -> DomTree {
-///     let b = true;
-///     todo!()
-/// }
-/// ```
-#[proc_macro_error::proc_macro_error]
 #[proc_macro]
-pub fn rsx(s: TokenStream) -> TokenStream {
-    match syn::parse::<rsx::CallBody>(s) {
+pub fn rsx(tokens: TokenStream) -> TokenStream {
+    match syn::parse::<rsx::CallBody>(tokens) {
         Err(err) => err.to_compile_error().into(),
-        Ok(stream) => stream.to_token_stream().into(),
+        Ok(body) => RenderCallBody(body).into_token_stream().into(),
     }
 }
 
-/// Derive macro used to mark an enum as Routable.
+/// The rsx! macro makes it easy for developers to write jsx-style markup in their components.
 ///
-/// This macro can only be used on enums. Every varient of the macro needs to be marked
-/// with the `at` attribute to specify the URL of the route. It generates an implementation of
-///  `yew_router::Routable` trait and `const`s for the routes passed which are used with `Route`
-/// component.
-///
-/// # Example
-///
-/// ```
-/// # use yew_router::Routable;
-/// #[derive(Debug, Clone, Copy, PartialEq, Routable)]
-/// enum Routes {
-///     #[at("/")]
-///     Home,
-///     #[at("/secure")]
-///     Secure,
-///     #[at("/profile/{id}")]
-///     Profile(u32),
-///     #[at("/404")]
-///     NotFound,
-/// }
-/// ```
-#[proc_macro_derive(Routable, attributes(at, not_found))]
-pub fn routable_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStream {
-    use router::{routable_derive_impl, Routable};
-    let input = parse_macro_input!(input as Routable);
-    routable_derive_impl(input).into()
+/// The render macro automatically renders rsx - making it unhygienic.
+#[deprecated(note = "Use `rsx!` instead.")]
+#[proc_macro]
+pub fn render(tokens: TokenStream) -> TokenStream {
+    rsx(tokens)
 }
 
 /// Derive props for a component within the component definition.
@@ -225,10 +64,10 @@ pub fn routable_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStrea
 /// you would be repeating a lot of the usual Rust boilerplate.
 ///
 /// # Example
-/// ```
+/// ```rust,ignore
 /// #[inline_props]
-/// fn app(cx: Scope, bob: String) -> Element {
-///     cx.render(rsx!("hello, {bob}"))
+/// fn app(bob: String) -> Element {
+///     rsx!("hello, {bob}"))
 /// }
 ///
 /// // is equivalent to
@@ -239,13 +78,122 @@ pub fn routable_derive(input: proc_macro::TokenStream) -> proc_macro::TokenStrea
 /// }
 ///
 /// fn app(cx: Scope<AppProps>) -> Element {
-///     cx.render(rsx!("hello, {bob}"))
+///     rsx!("hello, {bob}"))
 /// }
 /// ```
 #[proc_macro_attribute]
-pub fn inline_props(_args: proc_macro::TokenStream, s: TokenStream) -> TokenStream {
-    match syn::parse::<inlineprops::InlinePropsBody>(s) {
+#[deprecated(note = "Use `#[component]` instead.")]
+pub fn inline_props(_args: TokenStream, s: TokenStream) -> TokenStream {
+    let comp_body = parse_macro_input!(s as ComponentBody);
+
+    match comp_body.deserialize(InlinePropsDeserializerArgs {}) {
         Err(e) => e.to_compile_error().into(),
-        Ok(s) => s.to_token_stream().into(),
+        Ok(output) => output.to_token_stream().into(),
+    }
+}
+
+pub(crate) const COMPONENT_ARG_CASE_CHECK_OFF: &str = "no_case_check";
+
+/// Streamlines component creation.
+/// This is the recommended way of creating components,
+/// though you might want lower-level control with more advanced uses.
+///
+/// # Arguments
+/// * `no_case_check` - Doesn't enforce `PascalCase` on your component names.
+/// **This will be removed/deprecated in a future update in favor of a more complete Clippy-backed linting system.**
+/// The reasoning behind this is that Clippy allows more robust and powerful lints, whereas
+/// macros are extremely limited.
+///
+/// # Features
+/// This attribute:
+/// * Enforces that your component uses `PascalCase`.
+/// No warnings are generated for the `PascalCase`
+/// function name, but everything else will still raise a warning if it's incorrectly `PascalCase`.
+/// Does not disable warnings anywhere else, so if you, for example,
+/// accidentally don't use `snake_case`
+/// for a variable name in the function, the compiler will still warn you.
+/// * Automatically uses `#[inline_props]` if there's more than 1 parameter in the function.
+/// * Verifies the validity of your component.
+/// E.g. if it has a [`Scope`](dioxus_core::Scope) argument.
+/// Notes:
+///     * This doesn't work 100% of the time, because of macro limitations.
+///     * Provides helpful messages if your component is not correct.
+/// Possible bugs (please, report these!):
+///     * There might be bugs where it incorrectly *denies* validity.
+/// This is bad as it means that you can't use the attribute or you have to change the component.
+///     * There might be bugs where it incorrectly *confirms* validity.
+/// You will still know if the component is invalid once you use it,
+/// but the error might be less helpful.
+///
+/// # Examples
+/// * Without props:
+/// ```rust,ignore
+/// #[component]
+/// fn GreetBob() -> Element {
+///     rsx! { "hello, bob" }
+/// }
+///
+/// // is equivalent to
+///
+/// #[allow(non_snake_case)]
+/// fn GreetBob() -> Element {
+///     #[warn(non_snake_case)]
+///     #[inline(always)]
+///     fn __dx_inner_comp() -> Element {
+///         rsx! { "hello, bob" }
+///     }
+///     // There's no function call overhead since __dx_inner_comp has the #[inline(always)] attribute,
+///     // so don't worry about performance.
+///     __dx_inner_comp(cx)
+/// }
+/// ```
+/// * With props:
+/// ```rust,ignore
+/// #[component(no_case_check)]
+/// fn GreetPerson(person: String) -> Element {
+///     rsx! { "hello, {person}" }
+/// }
+///
+/// // is equivalent to
+///
+/// #[derive(Props, PartialEq)]
+/// #[allow(non_camel_case_types)]
+/// struct GreetPersonProps {
+///     person: String,
+/// }
+///
+/// #[allow(non_snake_case)]
+/// fn GreetPerson(props: GreetPersonProps>) -> Element {
+///     #[warn(non_snake_case)]
+///     #[inline(always)]
+///     fn __dx_inner_comp(props: GreetPersonProps>e) -> Element {
+///         let GreetPersonProps { person } = &cx.props;
+///         {
+///             rsx! { "hello, {person}" }
+///         }
+///     }
+///
+///     __dx_inner_comp(cx)
+/// }
+/// ```
+// TODO: Maybe add an option to input a custom component name through the args.
+//  I think that's unnecessary, but there might be some scenario where it could be useful.
+#[proc_macro_attribute]
+pub fn component(args: TokenStream, input: TokenStream) -> TokenStream {
+    let component_body = parse_macro_input!(input as ComponentBody);
+    let case_check = match Punctuated::<Path, Token![,]>::parse_terminated.parse(args) {
+        Err(e) => return e.to_compile_error().into(),
+        Ok(args) => {
+            if let Some(first) = args.first() {
+                !first.is_ident(COMPONENT_ARG_CASE_CHECK_OFF)
+            } else {
+                true
+            }
+        }
+    };
+
+    match component_body.deserialize(ComponentDeserializerArgs { case_check }) {
+        Err(e) => e.to_compile_error().into(),
+        Ok(output) => output.to_token_stream().into(),
     }
 }
