@@ -1,257 +1,512 @@
-//! Instructions returned by the VirtualDOM on how to modify the Real DOM.
-//!
-//! This module contains an internal API to generate these instructions.
-//!
-//! Beware that changing code in this module will break compatibility with
-//! interpreters for these types of DomEdits.
+use rustc_hash::FxHashSet;
 
-use crate::innerlude::*;
-use std::{any::Any, fmt::Debug};
+use crate::{arena::ElementId, AttributeValue, ScopeId, Template};
 
-/// ## Mutations
+/// Something that can handle the mutations that are generated by the diffing process and apply them to the Real DOM
 ///
-/// This method returns "mutations" - IE the necessary changes to get the RealDOM to match the VirtualDOM. It also
-/// includes a list of NodeRefs that need to be applied and effects that need to be triggered after the RealDOM has
-/// applied the edits.
+/// This object provides a bunch of important information for a renderer to use patch the Real Dom with the state of the
+/// VirtualDom. This includes the scopes that were modified, the templates that were discovered, and a list of changes
+/// in the form of a [`Mutation`].
+///
+/// These changes are specific to one subtree, so to patch multiple subtrees, you'd need to handle each set separately.
+///
+/// Templates, however, apply to all subtrees, not just target subtree.
 ///
 /// Mutations are the only link between the RealDOM and the VirtualDOM.
-pub struct Mutations<'a> {
-    pub edits: Vec<DomEdit<'a>>,
-    pub dirty_scopes: FxHashSet<ScopeId>,
-    pub refs: Vec<NodeRefMutation<'a>>,
-}
+pub trait WriteMutations {
+    /// Register a template with the renderer
+    fn register_template(&mut self, template: Template);
 
-impl Debug for Mutations<'_> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("Mutations")
-            .field("edits", &self.edits)
-            .field("noderefs", &self.refs)
-            .finish()
-    }
-}
+    /// Add these m children to the target element
+    ///
+    /// Id: The ID of the element being mounted to
+    /// M: The number of nodes on the stack to append to the target element
+    fn append_children(&mut self, id: ElementId, m: usize);
 
-/// A `DomEdit` represents a serialized form of the VirtualDom's trait-based API. This allows streaming edits across the
-/// network or through FFI boundaries.
-#[derive(Debug, PartialEq)]
-#[cfg_attr(
-    feature = "serialize",
-    derive(serde::Serialize, serde::Deserialize),
-    serde(tag = "type")
-)]
-pub enum DomEdit<'bump> {
-    PushRoot {
-        root: u64,
-    },
+    /// Assign the element at the given path the target ElementId.
+    ///
+    /// The path is in the form of a list of indices based on children. Templates cannot have more than 255 children per
+    /// element, hence the use of a single byte.
+    ///
+    /// Path: The path of the child of the topmost node on the stack. A path of `[]` represents the topmost node. A path of `[0]` represents the first child. `[0,1,2]` represents 1st child's 2nd child's 3rd child.
+    /// Id: The ID we're assigning to this element/placeholder. This will be used later to modify the element or replace it with another element.
+    fn assign_node_id(&mut self, path: &'static [u8], id: ElementId);
 
-    AppendChildren {
-        many: u32,
-    },
+    /// Create a placeholder in the DOM that we will use later.
+    ///
+    /// Dioxus currently requires the use of placeholders to maintain a re-entrance point for things like list diffing
+    ///
+    /// Id: The ID we're assigning to this element/placeholder. This will be used later to modify the element or replace it with another element.
+    fn create_placeholder(&mut self, id: ElementId);
 
-    // // save a possibly-fragment node as a template
-    // SaveAsTemplate {
-    //     many: u32,
-    // },
+    /// Create a node specifically for text with the given value
+    ///
+    /// Value: The text content of this text node
+    /// Id: The ID we're assigning to this specific text nodes. This will be used later to modify the element or replace it with another element.
+    fn create_text_node(&mut self, value: &str, id: ElementId);
 
-    // "Root" refers to the item directly
-    // it's a waste of an instruction to push the root directly
-    ReplaceWith {
-        root: u64,
-        m: u32,
-    },
-    InsertAfter {
-        root: u64,
-        n: u32,
-    },
-    InsertBefore {
-        root: u64,
-        n: u32,
-    },
-    Remove {
-        root: u64,
-    },
-    CreateTextNode {
-        text: &'bump str,
-        root: u64,
-    },
-    CreateElement {
-        tag: &'bump str,
-        root: u64,
-    },
-    CreateElementNs {
-        tag: &'bump str,
-        root: u64,
-        ns: &'static str,
-    },
-    CreatePlaceholder {
-        root: u64,
-    },
-    NewEventListener {
-        event_name: &'static str,
-        scope: ScopeId,
-        root: u64,
-    },
-    RemoveEventListener {
-        root: u64,
-        event: &'static str,
-    },
-    SetText {
-        root: u64,
-        text: &'bump str,
-    },
-    SetAttribute {
-        root: u64,
-        field: &'static str,
-        value: &'bump str,
-        ns: Option<&'bump str>,
-    },
-    RemoveAttribute {
-        root: u64,
-        name: &'static str,
-    },
-}
+    /// Hydrate an existing text node at the given path with the given text.
+    ///
+    /// Assign this text node the given ID since we will likely need to modify this text at a later point
+    ///
+    /// Path: The path of the child of the topmost node on the stack. A path of `[]` represents the topmost node. A path of `[0]` represents the first child. `[0,1,2]` represents 1st child's 2nd child's 3rd child.
+    /// Value: The value of the textnode that we want to set the placeholder with
+    /// Id: The ID we're assigning to this specific text nodes. This will be used later to modify the element or replace it with another element.
+    fn hydrate_text_node(&mut self, path: &'static [u8], value: &str, id: ElementId);
 
-use fxhash::FxHashSet;
-use DomEdit::*;
+    /// Load and clone an existing node from a template saved under that specific name
+    ///
+    /// Dioxus guarantees that the renderer will have already been provided the template.
+    /// When the template is picked up in the template list, it should be saved under its "name" - here, the name
+    ///
+    /// Name: The unique "name" of the template based on the template location. When paired with `rsx!`, this is autogenerated
+    /// Index: The index root we loading from the template. The template is stored as a list of nodes. This index represents the position of that root
+    /// Id: The ID we're assigning to this element being loaded from the template (This will be used later to move the element around in lists)
+    fn load_template(&mut self, name: &'static str, index: usize, id: ElementId);
 
-impl<'a> Mutations<'a> {
-    pub(crate) fn new() -> Self {
-        Self {
-            edits: Vec::new(),
-            refs: Vec::new(),
-            dirty_scopes: Default::default(),
-        }
-    }
+    /// Replace the target element (given by its ID) with the topmost m nodes on the stack
+    ///
+    /// id: The ID of the node we're going to replace with new nodes
+    /// m: The number of nodes on the stack to replace the target element with
+    fn replace_node_with(&mut self, id: ElementId, m: usize);
 
-    // Navigation
-    pub(crate) fn push_root(&mut self, root: ElementId) {
-        let id = root.as_u64();
-        self.edits.push(PushRoot { root: id });
-    }
+    /// Replace an existing element in the template at the given path with the m nodes on the stack
+    ///
+    /// Path: The path of the child of the topmost node on the stack. A path of `[]` represents the topmost node. A path of `[0]` represents the first child. `[0,1,2]` represents 1st child's 2nd child's 3rd child.
+    /// M: The number of nodes on the stack to replace the target element with
+    fn replace_placeholder_with_nodes(&mut self, path: &'static [u8], m: usize);
 
-    pub(crate) fn replace_with(&mut self, root: ElementId, m: u32) {
-        let root = root.as_u64();
-        self.edits.push(ReplaceWith { m, root });
-    }
+    /// Insert a number of nodes after a given node.
+    ///
+    /// Id: The ID of the node to insert after.
+    /// M: The number of nodes on the stack to insert after the target node.
+    fn insert_nodes_after(&mut self, id: ElementId, m: usize);
 
-    pub(crate) fn insert_after(&mut self, root: ElementId, n: u32) {
-        let root = root.as_u64();
-        self.edits.push(InsertAfter { n, root });
-    }
+    /// Insert a number of nodes before a given node.
+    ///
+    /// Id: The ID of the node to insert before.
+    /// M: The number of nodes on the stack to insert before the target node.
+    fn insert_nodes_before(&mut self, id: ElementId, m: usize);
 
-    pub(crate) fn insert_before(&mut self, root: ElementId, n: u32) {
-        let root = root.as_u64();
-        self.edits.push(InsertBefore { n, root });
-    }
-
-    pub(crate) fn append_children(&mut self, n: u32) {
-        self.edits.push(AppendChildren { many: n });
-    }
-
-    // Remove Nodes from the dom
-    pub(crate) fn remove(&mut self, id: u64) {
-        self.edits.push(Remove { root: id });
-    }
-
-    // Create
-    pub(crate) fn create_text_node(&mut self, text: &'a str, id: ElementId) {
-        let id = id.as_u64();
-        self.edits.push(CreateTextNode { text, root: id });
-    }
-
-    pub(crate) fn create_element(
+    /// Set the value of a node's attribute.
+    ///
+    /// Name: The name of the attribute to set.
+    /// NS: The (optional) namespace of the attribute. For instance, "style" is in the "style" namespace.
+    /// Value: The value of the attribute.
+    /// Id: The ID of the node to set the attribute of.
+    fn set_attribute(
         &mut self,
-        tag: &'static str,
+        name: &'static str,
         ns: Option<&'static str>,
+        value: &AttributeValue,
+        id: ElementId,
+    );
+
+    /// Set the text content of a node.
+    ///
+    /// Value: The textcontent of the node
+    /// Id: The ID of the node to set the textcontent of.
+    fn set_node_text(&mut self, value: &str, id: ElementId);
+
+    /// Create a new Event Listener.
+    ///
+    /// Name: The name of the event to listen for.
+    /// Id: The ID of the node to attach the listener to.
+    fn create_event_listener(&mut self, name: &'static str, id: ElementId);
+
+    /// Remove an existing Event Listener.
+    ///
+    /// Name: The name of the event to remove.
+    /// Id: The ID of the node to remove.
+    fn remove_event_listener(&mut self, name: &'static str, id: ElementId);
+
+    /// Remove a particular node from the DOM
+    ///
+    /// Id: The ID of the node to remove.
+    fn remove_node(&mut self, id: ElementId);
+
+    /// Push the given root node onto our stack.
+    ///
+    /// Id: The ID of the root node to push.
+    fn push_root(&mut self, id: ElementId);
+
+    /// Swap to a new subtree
+    fn swap_subtree(&mut self, _subtree_index: usize) {}
+
+    /// Mark a scope as dirty
+    fn mark_scope_dirty(&mut self, _scope_id: ScopeId) {}
+}
+
+/// A `Mutation` represents a single instruction for the renderer to use to modify the UI tree to match the state
+/// of the Dioxus VirtualDom.
+///
+/// These edits can be serialized and sent over the network or through any interface
+#[derive(Debug, PartialEq)]
+pub enum Mutation {
+    /// Add these m children to the target element
+    AppendChildren {
+        /// The ID of the element being mounted to
+        id: ElementId,
+
+        /// The number of nodes on the stack to append to the target element
+        m: usize,
+    },
+
+    /// Assign the element at the given path the target ElementId.
+    ///
+    /// The path is in the form of a list of indices based on children. Templates cannot have more than 255 children per
+    /// element, hence the use of a single byte.
+    ///
+    ///
+    AssignId {
+        /// The path of the child of the topmost node on the stack
+        ///
+        /// A path of `[]` represents the topmost node. A path of `[0]` represents the first child.
+        /// `[0,1,2]` represents 1st child's 2nd child's 3rd child.
+        path: &'static [u8],
+
+        /// The ID we're assigning to this element/placeholder.
+        ///
+        /// This will be used later to modify the element or replace it with another element.
+        id: ElementId,
+    },
+
+    /// Create a placeholder in the DOM that we will use later.
+    ///
+    /// Dioxus currently requires the use of placeholders to maintain a re-entrance point for things like list diffing
+    CreatePlaceholder {
+        /// The ID we're assigning to this element/placeholder.
+        ///
+        /// This will be used later to modify the element or replace it with another element.
+        id: ElementId,
+    },
+
+    /// Create a node specifically for text with the given value
+    CreateTextNode {
+        /// The text content of this text node
+        value: String,
+
+        /// The ID we're assigning to this specific text nodes
+        ///
+        /// This will be used later to modify the element or replace it with another element.
+        id: ElementId,
+    },
+
+    /// Hydrate an existing text node at the given path with the given text.
+    ///
+    /// Assign this text node the given ID since we will likely need to modify this text at a later point
+    HydrateText {
+        /// The path of the child of the topmost node on the stack
+        ///
+        /// A path of `[]` represents the topmost node. A path of `[0]` represents the first child.
+        /// `[0,1,2]` represents 1st child's 2nd child's 3rd child.
+        path: &'static [u8],
+
+        /// The value of the textnode that we want to set the placeholder with
+        value: String,
+
+        /// The ID we're assigning to this specific text nodes
+        ///
+        /// This will be used later to modify the element or replace it with another element.
+        id: ElementId,
+    },
+
+    /// Load and clone an existing node from a template saved under that specific name
+    ///
+    /// Dioxus guarantees that the renderer will have already been provided the template.
+    /// When the template is picked up in the template list, it should be saved under its "name" - here, the name
+    LoadTemplate {
+        /// The "name" of the template. When paired with `rsx!`, this is autogenerated
+        name: &'static str,
+
+        /// Which root are we loading from the template?
+        ///
+        /// The template is stored as a list of nodes. This index represents the position of that root
+        index: usize,
+
+        /// The ID we're assigning to this element being loaded from the template
+        ///
+        /// This will be used later to move the element around in lists
+        id: ElementId,
+    },
+
+    /// Replace the target element (given by its ID) with the topmost m nodes on the stack
+    ReplaceWith {
+        /// The ID of the node we're going to replace with
+        id: ElementId,
+
+        /// The number of nodes on the stack to replace the target element with
+        m: usize,
+    },
+
+    /// Replace an existing element in the template at the given path with the m nodes on the stack
+    ReplacePlaceholder {
+        /// The path of the child of the topmost node on the stack
+        ///
+        /// A path of `[]` represents the topmost node. A path of `[0]` represents the first child.
+        /// `[0,1,2]` represents 1st child's 2nd child's 3rd child.
+        path: &'static [u8],
+
+        /// The number of nodes on the stack to replace the target element with
+        m: usize,
+    },
+
+    /// Insert a number of nodes after a given node.
+    InsertAfter {
+        /// The ID of the node to insert after.
+        id: ElementId,
+
+        /// The number of nodes on the stack to insert after the target node.
+        m: usize,
+    },
+
+    /// Insert a number of nodes before a given node.
+    InsertBefore {
+        /// The ID of the node to insert before.
+        id: ElementId,
+
+        /// The number of nodes on the stack to insert before the target node.
+        m: usize,
+    },
+
+    /// Set the value of a node's attribute.
+    SetAttribute {
+        /// The name of the attribute to set.
+        name: &'static str,
+
+        /// The (optional) namespace of the attribute.
+        /// For instance, "style" is in the "style" namespace.
+        ns: Option<&'static str>,
+
+        /// The value of the attribute.
+        value: AttributeValue,
+
+        /// The ID of the node to set the attribute of.
+        id: ElementId,
+    },
+
+    /// Set the textcontent of a node.
+    SetText {
+        /// The textcontent of the node
+        value: String,
+
+        /// The ID of the node to set the textcontent of.
+        id: ElementId,
+    },
+
+    /// Create a new Event Listener.
+    NewEventListener {
+        /// The name of the event to listen for.
+        name: String,
+
+        /// The ID of the node to attach the listener to.
+        id: ElementId,
+    },
+
+    /// Remove an existing Event Listener.
+    RemoveEventListener {
+        /// The name of the event to remove.
+        name: String,
+
+        /// The ID of the node to remove.
+        id: ElementId,
+    },
+
+    /// Remove a particular node from the DOM
+    Remove {
+        /// The ID of the node to remove.
+        id: ElementId,
+    },
+
+    /// Push the given root node onto our stack.
+    PushRoot {
+        /// The ID of the root node to push.
+        id: ElementId,
+    },
+}
+
+/// A static list of mutations that can be applied to the DOM. Note: this list does not contain any `Any` attribute values
+#[derive(Debug, PartialEq, Default)]
+pub struct Mutations {
+    /// The list of Scopes that were diffed, created, and removed during the Diff process.
+    pub dirty_scopes: FxHashSet<ScopeId>,
+
+    /// Any templates encountered while diffing the DOM.
+    ///
+    /// These must be loaded into a cache before applying the edits
+    pub templates: Vec<Template>,
+
+    /// Any mutations required to patch the renderer to match the layout of the VirtualDom
+    pub edits: Vec<Mutation>,
+}
+
+impl Mutations {
+    /// Rewrites IDs to just be "template", so you can compare the mutations
+    ///
+    /// Used really only for testing
+    pub fn santize(mut self) -> Self {
+        for edit in self.edits.iter_mut() {
+            if let Mutation::LoadTemplate { name, .. } = edit {
+                *name = "template"
+            }
+        }
+
+        self
+    }
+}
+
+impl WriteMutations for Mutations {
+    fn register_template(&mut self, template: Template) {
+        self.templates.push(template)
+    }
+
+    fn append_children(&mut self, id: ElementId, m: usize) {
+        self.edits.push(Mutation::AppendChildren { id, m })
+    }
+
+    fn assign_node_id(&mut self, path: &'static [u8], id: ElementId) {
+        self.edits.push(Mutation::AssignId { path, id })
+    }
+
+    fn create_placeholder(&mut self, id: ElementId) {
+        self.edits.push(Mutation::CreatePlaceholder { id })
+    }
+
+    fn create_text_node(&mut self, value: &str, id: ElementId) {
+        self.edits.push(Mutation::CreateTextNode {
+            value: value.into(),
+            id,
+        })
+    }
+
+    fn hydrate_text_node(&mut self, path: &'static [u8], value: &str, id: ElementId) {
+        self.edits.push(Mutation::HydrateText {
+            path,
+            value: value.into(),
+            id,
+        })
+    }
+
+    fn load_template(&mut self, name: &'static str, index: usize, id: ElementId) {
+        self.edits.push(Mutation::LoadTemplate { name, index, id })
+    }
+
+    fn replace_node_with(&mut self, id: ElementId, m: usize) {
+        self.edits.push(Mutation::ReplaceWith { id, m })
+    }
+
+    fn replace_placeholder_with_nodes(&mut self, path: &'static [u8], m: usize) {
+        self.edits.push(Mutation::ReplacePlaceholder { path, m })
+    }
+
+    fn insert_nodes_after(&mut self, id: ElementId, m: usize) {
+        self.edits.push(Mutation::InsertAfter { id, m })
+    }
+
+    fn insert_nodes_before(&mut self, id: ElementId, m: usize) {
+        self.edits.push(Mutation::InsertBefore { id, m })
+    }
+
+    fn set_attribute(
+        &mut self,
+        name: &'static str,
+        ns: Option<&'static str>,
+        value: &AttributeValue,
         id: ElementId,
     ) {
-        let id = id.as_u64();
-        match ns {
-            Some(ns) => self.edits.push(CreateElementNs { root: id, ns, tag }),
-            None => self.edits.push(CreateElement { root: id, tag }),
-        }
-    }
-    // placeholders are nodes that don't get rendered but still exist as an "anchor" in the real dom
-    pub(crate) fn create_placeholder(&mut self, id: ElementId) {
-        let id = id.as_u64();
-        self.edits.push(CreatePlaceholder { root: id });
-    }
-
-    // events
-    pub(crate) fn new_event_listener(&mut self, listener: &Listener, scope: ScopeId) {
-        let Listener {
-            event,
-            mounted_node,
-            ..
-        } = listener;
-
-        let element_id = mounted_node.get().unwrap().as_u64();
-
-        self.edits.push(NewEventListener {
-            scope,
-            event_name: event,
-            root: element_id,
-        });
-    }
-    pub(crate) fn remove_event_listener(&mut self, event: &'static str, root: u64) {
-        self.edits.push(RemoveEventListener { event, root });
-    }
-
-    // modify
-    pub(crate) fn set_text(&mut self, text: &'a str, root: u64) {
-        self.edits.push(SetText { text, root });
-    }
-
-    pub(crate) fn set_attribute(&mut self, attribute: &'a Attribute, root: u64) {
-        let Attribute {
+        self.edits.push(Mutation::SetAttribute {
             name,
-            value,
-            namespace,
-            ..
-        } = attribute;
-
-        self.edits.push(SetAttribute {
-            field: name,
-            value,
-            ns: *namespace,
-            root,
-        });
+            ns,
+            value: match value {
+                AttributeValue::Text(s) => AttributeValue::Text(s.clone()),
+                AttributeValue::Bool(b) => AttributeValue::Bool(*b),
+                AttributeValue::Float(n) => AttributeValue::Float(*n),
+                AttributeValue::Int(n) => AttributeValue::Int(*n),
+                AttributeValue::None => AttributeValue::None,
+                _ => panic!("Cannot serialize attribute value"),
+            },
+            id,
+        })
     }
 
-    pub(crate) fn remove_attribute(&mut self, attribute: &Attribute, root: u64) {
-        let name = attribute.name;
-        self.edits.push(RemoveAttribute { name, root });
+    fn set_node_text(&mut self, value: &str, id: ElementId) {
+        self.edits.push(Mutation::SetText {
+            value: value.into(),
+            id,
+        })
     }
 
-    pub(crate) fn mark_dirty_scope(&mut self, scope: ScopeId) {
-        self.dirty_scopes.insert(scope);
+    fn create_event_listener(&mut self, name: &'static str, id: ElementId) {
+        self.edits.push(Mutation::NewEventListener {
+            name: name.into(),
+            id,
+        })
+    }
+
+    fn remove_event_listener(&mut self, name: &'static str, id: ElementId) {
+        self.edits.push(Mutation::RemoveEventListener {
+            name: name.into(),
+            id,
+        })
+    }
+
+    fn remove_node(&mut self, id: ElementId) {
+        self.edits.push(Mutation::Remove { id })
+    }
+
+    fn push_root(&mut self, id: ElementId) {
+        self.edits.push(Mutation::PushRoot { id })
+    }
+
+    fn swap_subtree(&mut self, _subtree_index: usize) {}
+
+    fn mark_scope_dirty(&mut self, scope_id: ScopeId) {
+        self.dirty_scopes.insert(scope_id);
     }
 }
 
-// refs are only assigned once
-pub struct NodeRefMutation<'a> {
-    pub element: &'a mut Option<once_cell::sync::OnceCell<Box<dyn Any>>>,
-    pub element_id: ElementId,
-}
+/// A struct that ignores all mutations
+pub struct NoOpMutations;
 
-impl<'a> std::fmt::Debug for NodeRefMutation<'a> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("NodeRefMutation")
-            .field("element_id", &self.element_id)
-            .finish()
-    }
-}
+impl WriteMutations for NoOpMutations {
+    fn register_template(&mut self, _: Template) {}
 
-impl<'a> NodeRefMutation<'a> {
-    pub fn downcast_ref<T: 'static>(&self) -> Option<&T> {
-        self.element
-            .as_ref()
-            .and_then(|f| f.get())
-            .and_then(|f| f.downcast_ref::<T>())
+    fn append_children(&mut self, _: ElementId, _: usize) {}
+
+    fn assign_node_id(&mut self, _: &'static [u8], _: ElementId) {}
+
+    fn create_placeholder(&mut self, _: ElementId) {}
+
+    fn create_text_node(&mut self, _: &str, _: ElementId) {}
+
+    fn hydrate_text_node(&mut self, _: &'static [u8], _: &str, _: ElementId) {}
+
+    fn load_template(&mut self, _: &'static str, _: usize, _: ElementId) {}
+
+    fn replace_node_with(&mut self, _: ElementId, _: usize) {}
+
+    fn replace_placeholder_with_nodes(&mut self, _: &'static [u8], _: usize) {}
+
+    fn insert_nodes_after(&mut self, _: ElementId, _: usize) {}
+
+    fn insert_nodes_before(&mut self, _: ElementId, _: usize) {}
+
+    fn set_attribute(
+        &mut self,
+        _: &'static str,
+        _: Option<&'static str>,
+        _: &AttributeValue,
+        _: ElementId,
+    ) {
     }
-    pub fn downcast_mut<T: 'static>(&mut self) -> Option<&mut T> {
-        self.element
-            .as_mut()
-            .and_then(|f| f.get_mut())
-            .and_then(|f| f.downcast_mut::<T>())
-    }
+
+    fn set_node_text(&mut self, _: &str, _: ElementId) {}
+
+    fn create_event_listener(&mut self, _: &'static str, _: ElementId) {}
+
+    fn remove_event_listener(&mut self, _: &'static str, _: ElementId) {}
+
+    fn remove_node(&mut self, _: ElementId) {}
+
+    fn push_root(&mut self, _: ElementId) {}
 }
