@@ -1,29 +1,138 @@
-use std::rc::Rc;
+use crate::{
+    app::SharedContext,
+    assets::AssetHandlerRegistry,
+    edits::EditQueue,
+    ipc::{EventData, UserWindowEvent},
+    query::QueryEngine,
+    shortcut::{HotKey, ShortcutHandle, ShortcutRegistryError},
+    webview::WebviewInstance,
+    AssetRequest, Config, WryEventHandler,
+};
+use dioxus_core::{
+    prelude::{current_scope_id, ScopeId},
+    VirtualDom,
+};
+use dioxus_interpreter_js::MutationState;
+use std::{cell::RefCell, rc::Rc, rc::Weak};
+use tao::{
+    event::Event,
+    event_loop::EventLoopWindowTarget,
+    window::{Fullscreen as WryFullscreen, Window, WindowId},
+};
+use wry::{RequestAsyncResponder, WebView};
 
-use dioxus_core::ScopeState;
-use wry::application::event_loop::EventLoopProxy;
+#[cfg(target_os = "ios")]
+use tao::platform::ios::WindowExtIOS;
 
-use crate::UserWindowEvent;
-
-type ProxyType = EventLoopProxy<UserWindowEvent>;
-
-/// Desktop-Window handle api context
+/// Get an imperative handle to the current window without using a hook
 ///
-/// you can use this context control some window event
+/// ## Panics
+///
+/// This function will panic if it is called outside of the context of a Dioxus App.
+pub fn window() -> DesktopContext {
+    dioxus_core::prelude::consume_context()
+}
+
+/// A handle to the [`DesktopService`] that can be passed around.
+pub type DesktopContext = Rc<DesktopService>;
+
+/// An imperative interface to the current window.
+///
+/// To get a handle to the current window, use the [`use_window`] hook.
+///
+///
+/// # Example
 ///
 /// you can use `cx.consume_context::<DesktopContext>` to get this context
 ///
-/// ```rust
+/// ```rust, ignore
 ///     let desktop = cx.consume_context::<DesktopContext>().unwrap();
 /// ```
-#[derive(Clone)]
-pub struct DesktopContext {
-    proxy: ProxyType,
+pub struct DesktopService {
+    /// The wry/tao proxy to the current window
+    pub webview: WebView,
+
+    /// The tao window itself
+    pub window: Window,
+
+    pub(crate) shared: Rc<SharedContext>,
+
+    /// The receiver for queries about the current window
+    pub(super) query: QueryEngine,
+    pub(crate) edit_queue: EditQueue,
+    pub(crate) mutation_state: RefCell<MutationState>,
+    pub(crate) asset_handlers: AssetHandlerRegistry,
+
+    #[cfg(target_os = "ios")]
+    pub(crate) views: Rc<RefCell<Vec<*mut objc::runtime::Object>>>,
 }
 
-impl DesktopContext {
-    pub(crate) fn new(proxy: ProxyType) -> Self {
-        Self { proxy }
+/// A smart pointer to the current window.
+impl std::ops::Deref for DesktopService {
+    type Target = Window;
+
+    fn deref(&self) -> &Self::Target {
+        &self.window
+    }
+}
+
+impl DesktopService {
+    pub(crate) fn new(
+        webview: WebView,
+        window: Window,
+        shared: Rc<SharedContext>,
+        edit_queue: EditQueue,
+        asset_handlers: AssetHandlerRegistry,
+    ) -> Self {
+        Self {
+            window,
+            webview,
+            shared,
+            edit_queue,
+            mutation_state: Default::default(),
+            asset_handlers,
+            query: Default::default(),
+            #[cfg(target_os = "ios")]
+            views: Default::default(),
+        }
+    }
+
+    /// Send a list of mutations to the webview
+    pub(crate) fn send_edits(&self) {
+        let mut mutations = self.mutation_state.borrow_mut();
+        let serialized_edits = mutations.export_memory();
+        self.edit_queue.add_edits(serialized_edits);
+    }
+
+    /// Create a new window using the props and window builder
+    ///
+    /// Returns the webview handle for the new window.
+    ///
+    /// You can use this to control other windows from the current window.
+    ///
+    /// Be careful to not create a cycle of windows, or you might leak memory.
+    pub fn new_window(&self, dom: VirtualDom, cfg: Config) -> Weak<DesktopService> {
+        let window = WebviewInstance::new(cfg, dom, self.shared.clone());
+
+        let cx = window.dom.in_runtime(|| {
+            ScopeId::ROOT
+                .consume_context::<Rc<DesktopService>>()
+                .unwrap()
+        });
+
+        self.shared
+            .proxy
+            .send_event(UserWindowEvent(EventData::NewWindow, cx.id()))
+            .unwrap();
+
+        self.shared
+            .proxy
+            .send_event(UserWindowEvent(EventData::Poll, cx.id()))
+            .unwrap();
+
+        self.shared.pending_webviews.borrow_mut().push(window);
+
+        Rc::downgrade(&cx)
     }
 
     /// trigger the drag-window event
@@ -31,85 +140,175 @@ impl DesktopContext {
     /// Moves the window with the left mouse button until the button is released.
     ///
     /// you need use it in `onmousedown` event:
-    /// ```rust
+    /// ```rust, ignore
     /// onmousedown: move |_| { desktop.drag_window(); }
     /// ```
     pub fn drag(&self) {
-        let _ = self.proxy.send_event(UserWindowEvent::DragWindow);
+        if self.window.fullscreen().is_none() {
+            _ = self.window.drag_window();
+        }
     }
 
-    /// set window minimize state
-    pub fn set_minimized(&self, minimized: bool) {
-        let _ = self.proxy.send_event(UserWindowEvent::Minimize(minimized));
+    /// Toggle whether the window is maximized or not
+    pub fn toggle_maximized(&self) {
+        self.window.set_maximized(!self.window.is_maximized())
     }
 
-    /// set window maximize state
-    pub fn set_maximized(&self, maximized: bool) {
-        let _ = self.proxy.send_event(UserWindowEvent::Maximize(maximized));
-    }
-
-    /// set window visible or not
-    pub fn set_visible(&self, visible: bool) {
-        let _ = self.proxy.send_event(UserWindowEvent::Visible(visible));
-    }
-
-    /// close window
+    /// Close this window
     pub fn close(&self) {
-        let _ = self.proxy.send_event(UserWindowEvent::CloseWindow);
+        let _ = self
+            .shared
+            .proxy
+            .send_event(UserWindowEvent(EventData::CloseWindow, self.id()));
     }
 
-    /// set window to focus
-    pub fn focus(&self) {
-        let _ = self.proxy.send_event(UserWindowEvent::FocusWindow);
+    /// Close a particular window, given its ID
+    pub fn close_window(&self, id: WindowId) {
+        let _ = self
+            .shared
+            .proxy
+            .send_event(UserWindowEvent(EventData::CloseWindow, id));
     }
 
     /// change window to fullscreen
     pub fn set_fullscreen(&self, fullscreen: bool) {
-        let _ = self
-            .proxy
-            .send_event(UserWindowEvent::Fullscreen(fullscreen));
+        if let Some(handle) = &self.window.current_monitor() {
+            self.window.set_fullscreen(
+                fullscreen.then_some(WryFullscreen::Borderless(Some(handle.clone()))),
+            );
+        }
     }
 
-    /// set resizable state
-    pub fn set_resizable(&self, resizable: bool) {
-        let _ = self.proxy.send_event(UserWindowEvent::Resizable(resizable));
+    /// launch print modal
+    pub fn print(&self) {
+        if let Err(e) = self.webview.print() {
+            tracing::warn!("Open print modal failed: {e}");
+        }
     }
 
-    /// set the window always on top
-    pub fn set_always_on_top(&self, top: bool) {
-        let _ = self.proxy.send_event(UserWindowEvent::AlwaysOnTop(top));
+    /// Set the zoom level of the webview
+    pub fn set_zoom_level(&self, level: f64) {
+        self.webview.zoom(level);
     }
 
-    // set cursor visible or not
-    pub fn set_cursor_visible(&self, visible: bool) {
-        let _ = self
-            .proxy
-            .send_event(UserWindowEvent::CursorVisible(visible));
+    /// opens DevTool window
+    pub fn devtool(&self) {
+        #[cfg(debug_assertions)]
+        self.webview.open_devtools();
+
+        #[cfg(not(debug_assertions))]
+        tracing::warn!("Devtools are disabled in release builds");
     }
 
-    // set cursor grab
-    pub fn set_cursor_grab(&self, grab: bool) {
-        let _ = self.proxy.send_event(UserWindowEvent::CursorGrab(grab));
+    /// Create a wry event handler that listens for wry events.
+    /// This event handler is scoped to the currently active window and will only recieve events that are either global or related to the current window.
+    ///
+    /// The id this function returns can be used to remove the event handler with [`DesktopContext::remove_wry_event_handler`]
+    pub fn create_wry_event_handler(
+        &self,
+        handler: impl FnMut(&Event<UserWindowEvent>, &EventLoopWindowTarget<UserWindowEvent>) + 'static,
+    ) -> WryEventHandler {
+        self.shared.event_handlers.add(self.window.id(), handler)
     }
 
-    /// set window title
-    pub fn set_title(&self, title: &str) {
-        let _ = self
-            .proxy
-            .send_event(UserWindowEvent::SetTitle(String::from(title)));
+    /// Remove a wry event handler created with [`DesktopContext::create_wry_event_handler`]
+    pub fn remove_wry_event_handler(&self, id: WryEventHandler) {
+        self.shared.event_handlers.remove(id)
     }
 
-    /// change window to borderless
-    pub fn set_decorations(&self, decoration: bool) {
-        let _ = self
-            .proxy
-            .send_event(UserWindowEvent::SetDecorations(decoration));
+    /// Create a global shortcut
+    ///
+    /// Linux: Only works on x11. See [this issue](https://github.com/tauri-apps/tao/issues/331) for more information.
+    pub fn create_shortcut(
+        &self,
+        hotkey: HotKey,
+        callback: impl FnMut() + 'static,
+    ) -> Result<ShortcutHandle, ShortcutRegistryError> {
+        self.shared
+            .shortcut_manager
+            .add_shortcut(hotkey, Box::new(callback))
+    }
+
+    /// Remove a global shortcut
+    pub fn remove_shortcut(&self, id: ShortcutHandle) {
+        self.shared.shortcut_manager.remove_shortcut(id)
+    }
+
+    /// Remove all global shortcuts
+    pub fn remove_all_shortcuts(&self) {
+        self.shared.shortcut_manager.remove_all()
+    }
+
+    /// Provide a callback to handle asset loading yourself.
+    /// If the ScopeId isn't provided, defaults to a global handler.
+    /// Note that the handler is namespaced by name, not ScopeId.
+    ///
+    /// When the component is dropped, the handler is removed.
+    ///
+    /// See [`use_asset_handle`](crate::use_asset_handle) for a convenient hook.
+    pub fn register_asset_handler(
+        &self,
+        name: String,
+        handler: Box<dyn Fn(AssetRequest, RequestAsyncResponder) + 'static>,
+        scope: Option<ScopeId>,
+    ) {
+        self.asset_handlers.register_handler(
+            name,
+            handler,
+            scope.unwrap_or(current_scope_id().unwrap_or(ScopeId(0))),
+        )
+    }
+
+    /// Removes an asset handler by its identifier.
+    ///
+    /// Returns `None` if the handler did not exist.
+    pub fn remove_asset_handler(&self, name: &str) -> Option<()> {
+        self.asset_handlers.remove_handler(name).map(|_| ())
+    }
+
+    /// Push an objc view to the window
+    #[cfg(target_os = "ios")]
+    pub fn push_view(&self, view: objc_id::ShareId<objc::runtime::Object>) {
+        let window = &self.window;
+
+        unsafe {
+            use objc::runtime::Object;
+            use objc::*;
+            assert!(is_main_thread());
+            let ui_view = window.ui_view() as *mut Object;
+            let ui_view_frame: *mut Object = msg_send![ui_view, frame];
+            let _: () = msg_send![view, setFrame: ui_view_frame];
+            let _: () = msg_send![view, setAutoresizingMask: 31];
+
+            let ui_view_controller = window.ui_view_controller() as *mut Object;
+            let _: () = msg_send![ui_view_controller, setView: view];
+            self.views.borrow_mut().push(ui_view);
+        }
+    }
+
+    /// Pop an objc view from the window
+    #[cfg(target_os = "ios")]
+    pub fn pop_view(&self) {
+        let window = &self.window;
+
+        unsafe {
+            use objc::runtime::Object;
+            use objc::*;
+            assert!(is_main_thread());
+            if let Some(view) = self.views.borrow_mut().pop() {
+                let ui_view_controller = window.ui_view_controller() as *mut Object;
+                let _: () = msg_send![ui_view_controller, setView: view];
+            }
+        }
     }
 }
 
-/// use this function can get the `DesktopContext` context.
-pub fn use_window(cx: &ScopeState) -> &Rc<DesktopContext> {
-    cx.use_hook(|_| cx.consume_context::<DesktopContext>())
-        .as_ref()
-        .unwrap()
+#[cfg(target_os = "ios")]
+fn is_main_thread() -> bool {
+    use objc::runtime::{Class, BOOL, NO};
+    use objc::*;
+
+    let cls = Class::get("NSThread").unwrap();
+    let result: BOOL = unsafe { msg_send![cls, isMainThread] };
+    result != NO
 }
