@@ -1,3 +1,7 @@
+#![doc(html_logo_url = "https://avatars.githubusercontent.com/u/79236386")]
+#![doc(html_favicon_url = "https://avatars.githubusercontent.com/u/79236386")]
+#![deny(missing_docs)]
+
 //! Dioxus WebSys
 //!
 //! ## Overview
@@ -7,9 +11,8 @@
 //! - idle work
 //! - animations
 //! - jank-free rendering
-//! - noderefs
 //! - controlled components
-//! - re-hydration
+//! - hydration
 //! - and more.
 //!
 //! The actual implementation is farily thin, with the heavy lifting happening inside the Dioxus Core crate.
@@ -24,7 +27,7 @@
 // main thread.
 //
 // React solves this problem by breaking up the rendering process into a "diff" phase and a "render" phase. In Dioxus,
-// the diff phase is non-blocking, using "yield_now" to allow the browser to process other events. When the diff phase
+// the diff phase is non-blocking, using "work_with_deadline" to allow the browser to process other events. When the diff phase
 // is  finally complete, the VirtualDOM will return a set of "Mutations" for this crate to apply.
 //
 // Here, we schedule the "diff" phase during the browser's idle period, achieved by calling RequestIdleCallback and then
@@ -54,105 +57,31 @@
 
 use std::rc::Rc;
 
-pub use crate::cfg::WebConfig;
-use dioxus::SchedulerMsg;
-use dioxus::VirtualDom;
-pub use dioxus_core as dioxus;
-use dioxus_core::prelude::Component;
-use futures_util::FutureExt;
+pub use crate::cfg::Config;
+#[cfg(feature = "file_engine")]
+pub use crate::file_engine::WebFileEngineExt;
+use dioxus_core::VirtualDom;
+use futures_util::{pin_mut, select, FutureExt, StreamExt};
 
-mod cache;
 mod cfg;
 mod dom;
+#[cfg(feature = "eval")]
+mod eval;
+mod event;
+pub mod launch;
+mod mutations;
+pub use event::*;
+#[cfg(feature = "file_engine")]
+mod file_engine;
+#[cfg(all(feature = "hot_reload", debug_assertions))]
+mod hot_reload;
+#[cfg(feature = "hydrate")]
 mod rehydrate;
-mod ric_raf;
 
-/// Launch the VirtualDOM given a root component and a configuration.
-///
-/// This function expects the root component to not have root props. To launch the root component with root props, use
-/// `launch_with_props` instead.
-///
-/// This method will block the thread with `spawn_local` from wasm_bindgen_futures.
-///
-/// If you need to run the VirtualDOM in its own thread, use `run_with_props` instead and await the future.
-///
-/// # Example
-///
-/// ```rust, ignore
-/// fn main() {
-///     dioxus_web::launch(App);
-/// }
-///
-/// static App: Component = |cx| {
-///     rsx!(cx, div {"hello world"})
-/// }
-/// ```
-pub fn launch(root_component: Component) {
-    launch_with_props(root_component, (), |c| c);
-}
-
-/// Launch your app and run the event loop, with configuration.
-///
-/// This function will start your web app on the main web thread.
-///
-/// You can configure the WebView window with a configuration closure
-///
-/// ```rust
-/// use dioxus::prelude::*;
-///
-/// fn main() {
-///     dioxus_web::launch_with_props(App, |config| config.pre_render(true));
-/// }
-///
-/// fn app(cx: Scope) -> Element {
-///     cx.render(rsx!{
-///         h1 {"hello world!"}
-///     })
-/// }
-/// ```
-pub fn launch_cfg(root: Component, config_builder: impl FnOnce(&mut WebConfig) -> &mut WebConfig) {
-    launch_with_props(root, (), config_builder)
-}
-
-/// Launches the VirtualDOM from the specified component function and props.
-///
-/// This method will block the thread with `spawn_local`
-///
-/// # Example
-///
-/// ```rust, ignore
-/// fn main() {
-///     dioxus_web::launch_with_props(
-///         App,
-///         RootProps { name: String::from("joe") },
-///         |config| config
-///     );
-/// }
-///
-/// #[derive(ParitalEq, Props)]
-/// struct RootProps {
-///     name: String
-/// }
-///
-/// static App: Component<RootProps> = |cx| {
-///     rsx!(cx, div {"hello {cx.props.name}"})
-/// }
-/// ```
-pub fn launch_with_props<T>(
-    root_component: Component<T>,
-    root_properties: T,
-    configuration_builder: impl FnOnce(&mut WebConfig) -> &mut WebConfig,
-) where
-    T: Send + 'static,
-{
-    if cfg!(feature = "panic_hook") {
-        console_error_panic_hook::set_once();
-    }
-
-    let mut config = WebConfig::default();
-    configuration_builder(&mut config);
-    wasm_bindgen_futures::spawn_local(run_with_props(root_component, root_properties, config));
-}
+// Currently disabled since it actually slows down immediate rendering
+// todo: only schedule non-immediate renders through ric/raf
+// mod ric_raf;
+// mod rehydrate;
 
 /// Runs the app as a future that can be scheduled around the main thread.
 ///
@@ -166,73 +95,118 @@ pub fn launch_with_props<T>(
 ///     wasm_bindgen_futures::spawn_local(app_fut);
 /// }
 /// ```
-pub async fn run_with_props<T: 'static + Send>(root: Component<T>, root_props: T, cfg: WebConfig) {
-    let mut dom = VirtualDom::new_with_props(root, root_props);
+pub async fn run(virtual_dom: VirtualDom, web_config: Config) {
+    tracing::info!("Starting up");
 
-    for s in crate::cache::BUILTIN_INTERNED_STRINGS {
-        wasm_bindgen::intern(s);
+    let mut dom = virtual_dom;
+
+    #[cfg(feature = "eval")]
+    {
+        // Eval
+        dom.in_runtime(|| {
+            eval::init_eval();
+        });
     }
-    for s in &cfg.cached_strings {
-        wasm_bindgen::intern(s);
+
+    #[cfg(feature = "panic_hook")]
+    if web_config.default_panic_hook {
+        console_error_panic_hook::set_once();
     }
 
-    let tasks = dom.get_scheduler_channel();
+    #[cfg(all(feature = "hot_reload", debug_assertions))]
+    let mut hotreload_rx = hot_reload::init();
 
-    let sender_callback: Rc<dyn Fn(SchedulerMsg)> =
-        Rc::new(move |event| tasks.unbounded_send(event).unwrap());
+    let (tx, mut rx) = futures_channel::mpsc::unbounded();
 
-    let should_hydrate = cfg.hydrate;
+    let should_hydrate = web_config.hydrate;
 
-    let mut websys_dom = dom::WebsysDom::new(cfg, sender_callback);
+    let mut websys_dom = dom::WebsysDom::new(web_config, tx);
 
-    log::trace!("rebuilding app");
+    tracing::info!("rebuilding app");
 
     if should_hydrate {
-        // todo: we need to split rebuild and initialize into two phases
-        // it's a waste to produce edits just to get the vdom loaded
-        let _ = dom.rebuild();
+        #[cfg(feature = "hydrate")]
+        {
+            dom.rebuild(&mut crate::rehydrate::OnlyWriteTemplates(&mut websys_dom));
 
-        if let Err(err) = websys_dom.rehydrate(&dom) {
-            log::error!(
-                "Rehydration failed {:?}. Rebuild DOM into element from scratch",
-                &err
-            );
+            if let Err(err) = websys_dom.rehydrate(&dom) {
+                tracing::error!("Rehydration failed. {:?}", err);
+                tracing::error!("Rebuild DOM into element from scratch");
+                websys_dom.root.set_text_content(None);
 
-            websys_dom.root.set_text_content(None);
+                dom.rebuild(&mut websys_dom);
 
-            // errrrr we should split rebuild into two phases
-            // one that initializes things and one that produces edits
-            let edits = dom.rebuild();
-
-            websys_dom.apply_edits(edits.edits);
+                websys_dom.flush_edits();
+            }
         }
     } else {
-        let edits = dom.rebuild();
-        websys_dom.apply_edits(edits.edits);
+        dom.rebuild(&mut websys_dom);
+
+        websys_dom.flush_edits();
     }
 
-    let work_loop = ric_raf::RafLoop::new();
+    // the mutations come back with nothing - we need to actually mount them
+    websys_dom.mount();
 
     loop {
-        log::trace!("waiting for work");
-        // if virtualdom has nothing, wait for it to have something before requesting idle time
-        // if there is work then this future resolves immediately.
-        dom.wait_for_work().await;
+        tracing::trace!("waiting for work");
 
-        log::trace!("working..");
+        // if virtual dom has nothing, wait for it to have something before requesting idle time
+        // if there is work then this future resolves immediately.
+        let (mut res, template) = {
+            let work = dom.wait_for_work().fuse();
+            pin_mut!(work);
+            let mut rx_next = rx.select_next_some();
+
+            #[cfg(all(feature = "hot_reload", debug_assertions))]
+            {
+                let mut hot_reload_next = hotreload_rx.select_next_some();
+                select! {
+                    _ = work => (None, None),
+                    new_template = hot_reload_next => (None, Some(new_template)),
+                    evt = rx_next => (Some(evt), None),
+                }
+            }
+            #[cfg(not(all(feature = "hot_reload", debug_assertions)))]
+            select! {
+                _ = work => (None, None),
+                evt = rx_next => (Some(evt), None),
+            }
+        };
+
+        if let Some(template) = template {
+            dom.replace_template(template);
+        }
+
+        // Dequeue all of the events from the channel in send order
+        // todo: we should re-order these if possible
+        while let Some(evt) = res {
+            dom.handle_event(
+                evt.name.as_str(),
+                Rc::new(evt.data),
+                evt.element,
+                evt.bubbles,
+            );
+            res = rx.try_next().transpose().unwrap().ok();
+        }
+
+        // Todo: This is currently disabled because it has a negative impact on response times for events but it could be re-enabled for tasks
+        // Jank free rendering
+        //
+        // 1. wait for the browser to give us "idle" time
+        // 2. During idle time, diff the dom
+        // 3. Stop diffing if the deadline is exceded
+        // 4. Wait for the animation frame to patch the dom
 
         // wait for the mainthread to schedule us in
-        let mut deadline = work_loop.wait_for_idle_time().await;
+        // let deadline = work_loop.wait_for_idle_time().await;
 
         // run the virtualdom work phase until the frame deadline is reached
-        let mutations = dom.work_with_deadline(|| (&mut deadline).now_or_never().is_some());
+        dom.render_immediate(&mut websys_dom);
 
         // wait for the animation frame to fire so we can apply our changes
-        work_loop.wait_for_raf().await;
+        // work_loop.wait_for_raf().await;
 
-        for edit in mutations {
-            // actually apply our changes during the animation frame
-            websys_dom.apply_edits(edit.edits);
-        }
+        websys_dom.flush_edits();
     }
 }
