@@ -1,169 +1,213 @@
 use crate::dom::WebsysDom;
-use dioxus_core::{VNode, VirtualDom};
-use wasm_bindgen::JsCast;
-use web_sys::{Comment, Element, Node, Text};
+use dioxus_core::prelude::*;
+use dioxus_core::AttributeValue;
+use dioxus_core::WriteMutations;
+use dioxus_core::{DynamicNode, ElementId, ScopeState, TemplateNode, VNode, VirtualDom};
+use dioxus_interpreter_js::save_template;
 
 #[derive(Debug)]
 pub enum RehydrationError {
-    NodeTypeMismatch,
-    NodeNotFound,
     VNodeNotInitialized,
 }
+
 use RehydrationError::*;
 
 impl WebsysDom {
     // we're streaming in patches, but the nodes already exist
     // so we're just going to write the correct IDs to the node and load them in
     pub fn rehydrate(&mut self, dom: &VirtualDom) -> Result<(), RehydrationError> {
-        let root = self
-            .root
-            .clone()
-            .dyn_into::<Node>()
-            .map_err(|_| NodeTypeMismatch)?;
-
         let root_scope = dom.base_scope();
-        let root_node = root_scope.root_node();
-
-        let mut nodes = vec![root];
-        let mut counter = vec![0];
-
-        let mut last_node_was_text = false;
+        let mut ids = Vec::new();
+        let mut to_mount = Vec::new();
 
         // Recursively rehydrate the dom from the VirtualDom
-        self.rehydrate_single(
-            &mut nodes,
-            &mut counter,
-            dom,
-            root_node,
-            &mut last_node_was_text,
-        )
+        self.rehydrate_scope(root_scope, dom, &mut ids, &mut to_mount)?;
+
+        dioxus_interpreter_js::hydrate(ids);
+
+        #[cfg(feature = "mounted")]
+        for id in to_mount {
+            self.send_mount_event(id);
+        }
+
+        Ok(())
     }
 
-    fn rehydrate_single(
+    fn rehydrate_scope(
         &mut self,
-        nodes: &mut Vec<Node>,
-        place: &mut Vec<u32>,
+        scope: &ScopeState,
         dom: &VirtualDom,
-        node: &VNode,
-        last_node_was_text: &mut bool,
+        ids: &mut Vec<u32>,
+        to_mount: &mut Vec<ElementId>,
     ) -> Result<(), RehydrationError> {
+        let vnode = scope.root_node();
+        self.rehydrate_vnode(dom, vnode, ids, to_mount)
+    }
+
+    fn rehydrate_vnode(
+        &mut self,
+        dom: &VirtualDom,
+        vnode: &VNode,
+        ids: &mut Vec<u32>,
+        to_mount: &mut Vec<ElementId>,
+    ) -> Result<(), RehydrationError> {
+        for (i, root) in vnode.template.get().roots.iter().enumerate() {
+            self.rehydrate_template_node(
+                dom,
+                vnode,
+                root,
+                ids,
+                to_mount,
+                Some(vnode.mounted_root(i, dom).ok_or(VNodeNotInitialized)?),
+            )?;
+        }
+        Ok(())
+    }
+
+    fn rehydrate_template_node(
+        &mut self,
+        dom: &VirtualDom,
+        vnode: &VNode,
+        node: &TemplateNode,
+        ids: &mut Vec<u32>,
+        to_mount: &mut Vec<ElementId>,
+        root_id: Option<ElementId>,
+    ) -> Result<(), RehydrationError> {
+        tracing::trace!("rehydrate template node: {:?}", node);
         match node {
-            VNode::Text(t) => {
-                let node_id = t.id.get().ok_or(VNodeNotInitialized)?;
-
-                let cur_place = place.last_mut().unwrap();
-
-                // skip over the comment element
-                if *last_node_was_text {
-                    if cfg!(debug_assertions) {
-                        let node = nodes.last().unwrap().child_nodes().get(*cur_place).unwrap();
-                        let node_text = node.dyn_into::<Comment>().unwrap();
-                        assert_eq!(node_text.data(), "spacer");
+            TemplateNode::Element {
+                children, attrs, ..
+            } => {
+                let mut mounted_id = root_id;
+                for attr in *attrs {
+                    if let dioxus_core::TemplateAttribute::Dynamic { id } = attr {
+                        let attributes = &*vnode.dynamic_attrs[*id];
+                        let id = vnode
+                            .mounted_dynamic_attribute(*id, dom)
+                            .ok_or(VNodeNotInitialized)?;
+                        for attribute in attributes {
+                            let value = &attribute.value;
+                            mounted_id = Some(id);
+                            if let AttributeValue::Listener(_) = value {
+                                if attribute.name == "onmounted" {
+                                    to_mount.push(id);
+                                }
+                            }
+                        }
                     }
-                    *cur_place += 1;
                 }
-
-                let node = nodes
-                    .last()
-                    .unwrap()
-                    .child_nodes()
-                    .get(*cur_place)
-                    .ok_or(NodeNotFound)?;
-
-                let _text_el = node.dyn_ref::<Text>().ok_or(NodeTypeMismatch)?;
-
-                // in debug we make sure the text is the same
-                if cfg!(debug_assertions) {
-                    let contents = _text_el.node_value().unwrap();
-                    assert_eq!(t.text, contents);
+                if let Some(id) = mounted_id {
+                    ids.push(id.0 as u32);
                 }
-
-                *last_node_was_text = true;
-
-                self.interpreter.set_node(node_id.0, node);
-                // self.nodes[node_id.0] = Some(node);
-
-                *cur_place += 1;
-            }
-
-            VNode::Element(vel) => {
-                let node_id = vel.id.get().ok_or(VNodeNotInitialized)?;
-
-                let cur_place = place.last_mut().unwrap();
-
-                let node = nodes.last().unwrap().child_nodes().get(*cur_place).unwrap();
-
-                use smallstr::SmallString;
-                use std::fmt::Write;
-
-                // 8 digits is enough, yes?
-                // 12 million nodes in one page?
-                let mut s: SmallString<[u8; 8]> = smallstr::SmallString::new();
-                write!(s, "{}", node_id).unwrap();
-
-                node.dyn_ref::<Element>()
-                    .unwrap()
-                    .set_attribute("dioxus-id", s.as_str())
-                    .unwrap();
-
-                self.interpreter.set_node(node_id.0, node.clone());
-                // self.nodes[node_id.0] = Some(node.clone());
-
-                *cur_place += 1;
-
-                nodes.push(node.clone());
-
-                place.push(0);
-
-                // we cant have the last node be text
-                let mut last_node_was_text = false;
-                for child in vel.children {
-                    self.rehydrate_single(nodes, place, dom, child, &mut last_node_was_text)?;
-                }
-
-                for listener in vel.listeners {
-                    self.interpreter.NewEventListener(
-                        listener.event,
-                        listener.mounted_node.get().unwrap().as_u64(),
-                        self.handler.as_ref().unchecked_ref(),
-                    );
-                }
-
-                place.pop();
-                nodes.pop();
-
-                if cfg!(debug_assertions) {
-                    let el = node.dyn_ref::<Element>().unwrap();
-                    let name = el.tag_name().to_lowercase();
-                    assert_eq!(name, vel.tag);
+                if !children.is_empty() {
+                    for child in *children {
+                        self.rehydrate_template_node(dom, vnode, child, ids, to_mount, None)?;
+                    }
                 }
             }
+            TemplateNode::Dynamic { id } | TemplateNode::DynamicText { id } => self
+                .rehydrate_dynamic_node(
+                    dom,
+                    &vnode.dynamic_nodes[*id],
+                    *id,
+                    vnode,
+                    ids,
+                    to_mount,
+                )?,
+            _ => {}
+        }
+        Ok(())
+    }
 
-            VNode::Placeholder(el) => {
-                let node_id = el.id.get().ok_or(VNodeNotInitialized)?;
-
-                let cur_place = place.last_mut().unwrap();
-                let node = nodes.last().unwrap().child_nodes().get(*cur_place).unwrap();
-
-                self.interpreter.set_node(node_id.0, node);
-
-                // self.nodes[node_id.0] = Some(node);
-
-                *cur_place += 1;
+    fn rehydrate_dynamic_node(
+        &mut self,
+        dom: &VirtualDom,
+        dynamic: &DynamicNode,
+        dynamic_node_index: usize,
+        vnode: &VNode,
+        ids: &mut Vec<u32>,
+        to_mount: &mut Vec<ElementId>,
+    ) -> Result<(), RehydrationError> {
+        tracing::trace!("rehydrate dynamic node: {:?}", dynamic);
+        match dynamic {
+            dioxus_core::DynamicNode::Text(_) | dioxus_core::DynamicNode::Placeholder(_) => {
+                ids.push(
+                    vnode
+                        .mounted_dynamic_node(dynamic_node_index, dom)
+                        .ok_or(VNodeNotInitialized)?
+                        .0 as u32,
+                );
             }
-
-            VNode::Fragment(el) => {
-                for el in el.children {
-                    self.rehydrate_single(nodes, place, dom, el, last_node_was_text)?;
+            dioxus_core::DynamicNode::Component(comp) => {
+                let scope = comp
+                    .mounted_scope(dynamic_node_index, vnode, dom)
+                    .ok_or(VNodeNotInitialized)?;
+                self.rehydrate_scope(scope, dom, ids, to_mount)?;
+            }
+            dioxus_core::DynamicNode::Fragment(fragment) => {
+                for vnode in fragment {
+                    self.rehydrate_vnode(dom, vnode, ids, to_mount)?;
                 }
-            }
-
-            VNode::Component(el) => {
-                let scope = dom.get_scope(el.scope.get().unwrap()).unwrap();
-                let node = scope.root_node();
-                self.rehydrate_single(nodes, place, dom, node, last_node_was_text)?;
             }
         }
         Ok(())
     }
+}
+
+/// During rehydration, we don't want to actually write anything to the DOM, but we do need to store any templates that were created. This struct is used to only write templates to the DOM.
+pub(crate) struct OnlyWriteTemplates<'a>(pub &'a mut WebsysDom);
+
+impl WriteMutations for OnlyWriteTemplates<'_> {
+    fn register_template(&mut self, template: Template) {
+        let mut roots = vec![];
+
+        for root in template.roots {
+            roots.push(self.0.create_template_node(root))
+        }
+
+        self.0
+            .templates
+            .insert(template.name.to_owned(), self.0.max_template_id);
+        save_template(roots, self.0.max_template_id);
+        self.0.max_template_id += 1
+    }
+
+    fn append_children(&mut self, _: ElementId, _: usize) {}
+
+    fn assign_node_id(&mut self, _: &'static [u8], _: ElementId) {}
+
+    fn create_placeholder(&mut self, _: ElementId) {}
+
+    fn create_text_node(&mut self, _: &str, _: ElementId) {}
+
+    fn hydrate_text_node(&mut self, _: &'static [u8], _: &str, _: ElementId) {}
+
+    fn load_template(&mut self, _: &'static str, _: usize, _: ElementId) {}
+
+    fn replace_node_with(&mut self, _: ElementId, _: usize) {}
+
+    fn replace_placeholder_with_nodes(&mut self, _: &'static [u8], _: usize) {}
+
+    fn insert_nodes_after(&mut self, _: ElementId, _: usize) {}
+
+    fn insert_nodes_before(&mut self, _: ElementId, _: usize) {}
+
+    fn set_attribute(
+        &mut self,
+        _: &'static str,
+        _: Option<&'static str>,
+        _: &AttributeValue,
+        _: ElementId,
+    ) {
+    }
+
+    fn set_node_text(&mut self, _: &str, _: ElementId) {}
+
+    fn create_event_listener(&mut self, _: &'static str, _: ElementId) {}
+
+    fn remove_event_listener(&mut self, _: &'static str, _: ElementId) {}
+
+    fn remove_node(&mut self, _: ElementId) {}
+
+    fn push_root(&mut self, _: ElementId) {}
 }
